% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A128}
\alias{A128}
\title{A128 example matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 2 columns.
}
\usage{
A128
}
\description{
A128 example matrix
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
