% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{addedByResearcher313}
\alias{addedByResearcher313}
\title{addedByResearcher313 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
addedByResearcher313
}
\description{
addedByResearcher313 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
