% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ageM18}
\alias{ageM18}
\title{ageM18 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ageM18
}
\description{
ageM18 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
