% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ageM2}
\alias{ageM2}
\title{ageM2 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ageM2
}
\description{
ageM2 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
