% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ageM313}
\alias{ageM313}
\title{ageM313 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ageM313
}
\description{
ageM313 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
