% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ageM32}
\alias{ageM32}
\title{ageM32 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ageM32
}
\description{
ageM32 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
