% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ageSD1}
\alias{ageSD1}
\title{ageSD1 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ageSD1
}
\description{
ageSD1 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
