% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ageSD18}
\alias{ageSD18}
\title{ageSD18 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ageSD18
}
\description{
ageSD18 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
