% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ageSD3}
\alias{ageSD3}
\title{ageSD3 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ageSD3
}
\description{
ageSD3 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
