% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ageSD313}
\alias{ageSD313}
\title{ageSD313 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ageSD313
}
\description{
ageSD313 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
