% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ageSD32}
\alias{ageSD32}
\title{ageSD32 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ageSD32
}
\description{
ageSD32 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
