% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ageSD4}
\alias{ageSD4}
\title{ageSD4 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ageSD4
}
\description{
ageSD4 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
