% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alphas128}
\alias{alphas128}
\title{alphas128 example vector}
\format{
An object of class \code{numeric} of length 9.
}
\usage{
alphas128
}
\description{
alphas128 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
