% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{burnout1}
\alias{burnout1}
\title{burnout1 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
burnout1
}
\description{
burnout1 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
