% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{burnout18}
\alias{burnout18}
\title{burnout18 example vector}
\format{
An object of class \code{character} of length 2.
}
\usage{
burnout18
}
\description{
burnout18 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
