% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{burnout2}
\alias{burnout2}
\title{burnout2 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
burnout2
}
\description{
burnout2 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
