% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{burnout201}
\alias{burnout201}
\title{burnout201 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
burnout201
}
\description{
burnout201 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
