% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{burnout32}
\alias{burnout32}
\title{burnout32 example vector}
\format{
An object of class \code{character} of length 2.
}
\usage{
burnout32
}
\description{
burnout32 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
