% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{burnout4}
\alias{burnout4}
\title{burnout4 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
burnout4
}
\description{
burnout4 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
