% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoTiMAcombinePseudoRawData.R
\name{combinePseudoRawData}
\alias{combinePseudoRawData}
\title{Title}
\usage{
combinePseudoRawData(listOfStudyFits = NULL, moderatorValues = NULL)
}
\arguments{
\item{listOfStudyFits}{?}

\item{moderatorValues}{?}
}
\value{

}
\description{
Title
}
\examples{

}
