% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combineVariablesNames128}
\alias{combineVariablesNames128}
\title{combineVariablesNames128 example vector}
\format{
An object of class \code{character} of length 3.
}
\usage{
combineVariablesNames128
}
\description{
combineVariablesNames128 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
