% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoTiMACompileListOfPrimaryStudies.R
\name{compileListOfPrimaryStudies}
\alias{compileListOfPrimaryStudies}
\title{compileListOfPrimaryStudies}
\usage{
compileListOfPrimaryStudies(selectedStudies = NULL, excludedElements = NULL)
}
\arguments{
\item{selectedStudies}{?}

\item{excludedElements}{?}
}
\value{

}
\description{
compileListOfPrimaryStudies
}
\examples{

}
