% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoTiMA.R
\name{CoTiMA}
\alias{CoTiMA}
\title{Title}
\usage{
CoTiMA(
  workingDirectory = NULL,
  sourceDirectory = NULL,
  resultsFileName = "CoTiMA.txt",
  filePrefix = "CoTiMA",
  primaryStudies = NULL,
  activateRPB = FALSE,
  checkSingleStudyResults = TRUE,
  nlatents = NULL,
  manifesttraitvar = FALSE,
  saveRawData = list(),
  retryattempts = 30,
  refits = 5,
  extraRefits = c(),
  factorExtraRefits = 1,
  NPSOL = FALSE,
  coresToUse = c(1),
  useCTMultigroupFitAlt = FALSE,
  plotCrossEffects = TRUE,
  plotAutoEffects = TRUE,
  timeUnit = "timeUnit",
  timeRange = c(),
  yLimitsForEffects = c(),
  digits = 4,
  testHeterogeneityModel = FALSE,
  testDRIFTallModel = TRUE,
  testDRIFTallWOdtModel = FALSE,
  testDRIFTSingleModel = FALSE,
  testDRIFTCrossModel = FALSE,
  testDRIFTAutoModel = FALSE,
  testAllInvariantModel = FALSE,
  testModeratorModel = FALSE,
  moderatorNumber = c(),
  testUserSpecifiedModel = list(singleStudyModelFits = NULL, moderatorValues = NULL,
    listOfFixedModelMatrices = list(T0VAR = NULL, DIFFUSION = NULL, DRIFT = NULL),
    listOfModeratorModelMatrices = list(T0VAR = NULL, DIFFUSION = NULL, DRIFT = NULL)),
  confidenceIntervals = FALSE,
  statisticalPower = c(),
  statisticalPowerRefits = 1,
  yLimitsForPower = c(),
  fixedAndRandomEffects = TRUE,
  publicationBias = FALSE,
  saveSingleStudyModelFit = c(),
  loadSingleStudyModelFit = c(),
  saveHeterogeneityModelFit = c(),
  loadHeterogeneityModelFit = c(),
  saveDRIFTAllModelFit = c(),
  loadDRIFTAllModelFit = c(),
  saveDRIFTSingleModelFit = c(),
  loadDRIFTSingleModelFit = c(),
  saveDRIFTCrossModelFit = c(),
  loadDRIFTCrossModelFit = c(),
  saveDRIFTAutoModelFit = c(),
  loadDRIFTAutoModelFit = c(),
  saveDRIFTallWOdtModelFit = c(),
  loadDRIFTallWOdtModelFit = c(),
  saveAllInvariantModelFit = c(),
  loadAllInvariantModelFit = c(),
  saveHomAllWOSingleDriftModelFit = c(),
  loadHomAllWOSingleDriftModelFit = c(),
  saveModeratorModelFit = c(),
  loadModeratorModelFit = c(),
  saveUserSpecifiedModel = c(),
  loadUserSpecifiedModel = c(),
  loadRawData = list(),
  saveSingleStudyFit = c(),
  loadSingleStudyFit = c(),
  saveHomogeneityStudyFit = FALSE,
  loadHomogeneityStudyFit = FALSE,
  saveHeterogeneityStudyFit = FALSE,
  loadHeterogeneityStudyFit = FALSE,
  loadFilePrefix = NULL,
  unixLike = FALSE,
  figureFilePrefix = c()
)
}
\arguments{
\item{workingDirectory}{workingDirectory}

\item{sourceDirectory}{sourceDirectory}

\item{resultsFileName}{the prefix for the result file that is created}

\item{filePrefix}{the prefix for the figure files and the model fit files that are created}

\item{primaryStudies}{list of lists: list(deltas, sampleSizes, empcovs, moderators, startValues, studyNumbers)}

\item{activateRPB}{set to TRUE to receive push messages with CoTiMA notifications on your phone}

\item{checkSingleStudyResults}{displays estimates from single study ctsem models and waits for user inoput to continue}

\item{nlatents}{n latent variables}

\item{manifesttraitvar}{test model with unit effects/unobserved heterogeneity (requires more then 2 time points could still fail then)}

\item{saveRawData}{TBN}

\item{retryattempts}{number of iterations to be used by ctsem}

\item{refits}{number of re-fits used by this CoTiMA Function}

\item{extraRefits}{(vector of) study number(s) that should be re-fitted even more (difficult models)}

\item{factorExtraRefits}{factor by which "refits" are multiplied for difficult models}

\item{NPSOL}{request NPSOL optimizer (sometimes better estimates, e.g., for confidence intervals)}

\item{coresToUse}{could be set to -n (usually -1) which is subtracted from the overall number available (on non Windows machines)}

\item{useCTMultigroupFitAlt}{TBN}

\item{plotCrossEffects}{plot cross effects across a range of discrete intervals}

\item{plotAutoEffects}{plot auto effects across a range of discrete intervals}

\item{timeUnit}{timelag unit to lable x-axis of plots}

\item{timeRange}{used for Plotting and Poweranalysis. Populated by 0 to 1.5*maxDelta (Steptwidth = 1) if not specified as list(min,max,stepWidth)}

\item{yLimitsForEffects}{used the y-axis of Drift-Plots. Populated by c("round(min(effects)-.05, 1)", "round(max(effects)-.05, 1)")}

\item{digits}{rounding used in output}

\item{testHeterogeneityModel}{multigoup model with all parameters freely estimated across groups}

\item{testDRIFTallModel}{multigoup model (CoTiMA!) with all DRIFT parameters invariantly estimated across groups}

\item{testDRIFTallWOdtModel}{TBN}

\item{testDRIFTSingleModel}{multigoup models with every single DRIFT parameters invariantly estimated across groups}

\item{testDRIFTCrossModel}{multigoup model to test equality of cross effects}

\item{testDRIFTAutoModel}{multigoup model to test equality of auto effects}

\item{testAllInvariantModel}{multigoup model (CoTiMA!) with ALL DRIFT paramters invariant. Required for Statistical Power Computations}

\item{testModeratorModel}{performs moderator analysis. Fails if no (vector of) moderator values are provided (e.g. moderator1 <- c(13))}

\item{moderatorNumber}{selects from the vector of moderator values (the 1st is used if not otehrwiseotherwise specificed)}

\item{testUserSpecifiedModel}{analysis of a user specified model}

\item{confidenceIntervals}{estimate confidence intervals (for all requested models)}

\item{statisticalPower}{beta, e.g., 95. Computes required sample sizes for discrete time drift paramters across range of time lags. Sets testAllInvariantModel=TRUE}

\item{statisticalPowerRefits}{TBN}

\item{yLimitsForPower}{used the y-axis of RequiredSampleSize-Plots. Populated by c("round(min(effects)-.05, 1)", "round(max(effects)-.05, 1)")}

\item{fixedAndRandomEffects}{performs fixed and random effects analysis}

\item{publicationBias}{perform Egger's test, funnel plots and PET-PEESE estimates}

\item{saveSingleStudyModelFit}{save the fit of single study ctsem models (could save a lot of time afterwards if the fit is loaded)}

\item{loadSingleStudyModelFit}{load the fit of single study ctsem models}

\item{saveHeterogeneityModelFit}{save the fit of the heterogeneity model}

\item{loadHeterogeneityModelFit}{load the fit of the heterogeneity model}

\item{saveDRIFTAllModelFit}{save the fit of the full CoTiMA model}

\item{loadDRIFTAllModelFit}{load the fit of the full CoTiMA model}

\item{saveDRIFTSingleModelFit}{save the fit of the CoTiMA models with single invariant Drift Parameters}

\item{loadDRIFTSingleModelFit}{load the fit of the CoTiMA models with single invariant Drift Parameters}

\item{saveDRIFTCrossModelFit}{TBN}

\item{loadDRIFTCrossModelFit}{TBN}

\item{saveDRIFTAutoModelFit}{TBN}

\item{loadDRIFTAutoModelFit}{TBN}

\item{saveDRIFTallWOdtModelFit}{TBN}

\item{loadDRIFTallWOdtModelFit}{TBN}

\item{saveAllInvariantModelFit}{TBN}

\item{loadAllInvariantModelFit}{TBN}

\item{saveHomAllWOSingleDriftModelFit}{save models required for calculation of statistical power (only relevant is statisticalPower is requested)}

\item{loadHomAllWOSingleDriftModelFit}{load models required for calculation of statistical power (only relevant is statisticalPower is requested)}

\item{saveModeratorModelFit}{save moderator models (the moderator number will be added)}

\item{loadModeratorModelFit}{load moderator models}

\item{saveUserSpecifiedModel}{TBN}

\item{loadUserSpecifiedModel}{TBN}

\item{loadRawData}{DEPRECATED: information which raw data to load is now included in list of primary studies}

\item{saveSingleStudyFit}{DEPRECATED: save the fit of single study ctsem models (could save a lot of time afterwards if the fit is loaded)}

\item{loadSingleStudyFit}{DEPRECATED: load the fit of single study ctsem models}

\item{saveHomogeneityStudyFit}{DEPRECATED: save the fit of the CoTiMA model}

\item{loadHomogeneityStudyFit}{load the fit of the CoTiMA model}

\item{saveHeterogeneityStudyFit}{save the fit of the heterogeneity model}

\item{loadHeterogeneityStudyFit}{load the fit of the heterogeneity model}

\item{loadFilePrefix}{DEPRECATED: the prefix of (older) files that contain model fits to be be loaded (useful if figureFilePrefix has changed since then)}

\item{unixLike}{DEPRECATED: instead the machine type is detected and parallelisation of loops with mclapply on unix machine activated}

\item{figureFilePrefix}{TBN}
}
\value{

}
\description{
Title
}
