% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoTiMABiG_D_BO}
\alias{CoTiMABiG_D_BO}
\title{ctmaBiG-object reproducing results of Guthier et al. (2020)}
\format{
An object of class \code{CoTiMAFit} of length 10.
}
\usage{
CoTiMABiG_D_BO
}
\description{
ctmaBiG-object reproducing results of Guthier et al. (2020)
}
\author{
C. Guthier, C. Dormann & J. Cortina \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
