% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoTiMAfittingSSMF.R
\name{CoTiMAfittingSSMF}
\alias{CoTiMAfittingSSMF}
\title{CoTiMAfittingSSMF}
\usage{
CoTiMAfittingSSMF(
  coresToUse,
  empraw,
  currentModel,
  refits,
  retryattempts,
  singleModelStartValues
)
}
\arguments{
\item{coresToUse}{?}

\item{empraw}{?}

\item{currentModel}{?}

\item{refits}{?}

\item{retryattempts}{?}

\item{singleModelStartValues}{?}
}
\value{

}
\description{
CoTiMAfittingSSMF
}
\examples{

}
