% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoTiMAFullFit1}
\alias{CoTiMAFullFit1}
\title{ctmaFit-object with a 'full' CoTiMA of 3 studies}
\format{
An object of class \code{CoTiMAFit} of length 20.
}
\usage{
CoTiMAFullFit1
}
\description{
ctmaFit-object with a 'full' CoTiMA of 3 studies
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
