% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoTiMAFullFit_6}
\alias{CoTiMAFullFit_6}
\title{ctmaFit-object with a 'full' CoTiMA of 6 studies}
\format{
An object of class \code{CoTiMAFit} of length 10.
}
\usage{
CoTiMAFullFit_6
}
\description{
ctmaFit-object with a 'full' CoTiMA of 6 studies
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
