% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoTiMAFullInv23Fit_6}
\alias{CoTiMAFullInv23Fit_6}
\title{1st fitted ctmaFit-object in a series of 2 to test equality of 2 cross effects}
\format{
An object of class \code{CoTiMAFit} of length 12.
}
\usage{
CoTiMAFullInv23Fit_6
}
\description{
1st fitted ctmaFit-object in a series of 2 to test equality of 2 cross effects
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
