% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoTiMAInitFit_6}
\alias{CoTiMAInitFit_6}
\title{ctmaInit-object with 6 primary studies}
\format{
An object of class \code{CoTiMAFit} of length 18.
}
\usage{
CoTiMAInitFit_6
}
\description{
ctmaInit-object with 6 primary studies
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
