% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoTiMAInitFit_6_NUTS}
\alias{CoTiMAInitFit_6_NUTS}
\title{ctmaInit-object with a 'full' CoTiMA of 6 studies using NUTS sampler}
\format{
An object of class \code{CoTiMAFit} of length 16.
}
\usage{
CoTiMAInitFit_6_NUTS
}
\description{
ctmaInit-object with a 'full' CoTiMA of 6 studies using NUTS sampler
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
