% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoTiMAInitFit_D_BO}
\alias{CoTiMAInitFit_D_BO}
\title{ctmaInit-object created by Guthier et al. (2020) with 48 primary studies}
\format{
An object of class \code{CoTiMAFit} of length 12.
}
\usage{
CoTiMAInitFit_D_BO
}
\description{
ctmaInit-object created by Guthier et al. (2020) with 48 primary studies
}
\author{
C. Guthier, C. Dormann & J. Cortina \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
