% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoTiMAMod1onFullFit_6}
\alias{CoTiMAMod1onFullFit_6}
\title{ctmaFit-object with a categorical moderator of the full drift matrix}
\format{
An object of class \code{CoTiMAFit} of length 13.
}
\usage{
CoTiMAMod1onFullFit_6
}
\description{
ctmaFit-object with a categorical moderator of the full drift matrix
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
