% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoTiMAMod2on23Fit_6}
\alias{CoTiMAMod2on23Fit_6}
\title{ctmaFit-object with a continuous moderator of 2 cross effects}
\format{
An object of class \code{CoTiMAFit} of length 15.
}
\usage{
CoTiMAMod2on23Fit_6
}
\description{
ctmaFit-object with a continuous moderator of 2 cross effects
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
