% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoTiMAPart134Inv3Fit_6}
\alias{CoTiMAPart134Inv3Fit_6}
\title{ctmaFit-object with with only one cross effect and this one set equal across primary studies}
\format{
An object of class \code{CoTiMAFit} of length 16.
}
\usage{
CoTiMAPart134Inv3Fit_6
}
\description{
ctmaFit-object with with only one cross effect and this one set equal across primary studies
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
