% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoTiMAPower_D_BO}
\alias{CoTiMAPower_D_BO}
\title{ctmaPower-object reproducing results of Guthier et al. (2020)}
\format{
An object of class \code{CoTiMAFit} of length 10.
}
\usage{
CoTiMAPower_D_BO
}
\description{
ctmaPower-object reproducing results of Guthier et al. (2020)
}
\author{
C. Guthier, C. Dormann & J. Cortina  \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
