% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoTiMAprep 1_0_0_3.R
\name{CoTiMAprep}
\alias{CoTiMAprep}
\title{CoTiMAprep}
\usage{
CoTiMAprep(
  activeDirectory = NULL,
  sourceDirectory = NULL,
  resultsFilePrefix = "CoTiMAprep",
  saveFilePrefix = "CoTiMAprep",
  loadFilePrefix = "CoTiMAprep",
  primaryStudies = NULL,
  activateRPB = FALSE,
  checkSingleStudyResults = TRUE,
  digits = 4,
  n.latent = NULL,
  loadRawData = list(),
  saveRawData = list(),
  retryattempts = 30,
  refits = 5,
  extraRefits = c(),
  factorExtraRefits = 1,
  NPSOL = FALSE,
  coresToUse = c(1),
  tryHard = TRUE,
  confidenceIntervals = FALSE,
  silentOverwrite = FALSE,
  saveSingleStudyModelFit = c(),
  loadSingleStudyModelFit = c()
)
}
\arguments{
\item{loadSingleStudyModelFit}{}
}
\value{

}
\description{
CoTiMAprep
}
\examples{

}
