% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoTiMASaveFile.R
\name{CoTiMASaveFile}
\alias{CoTiMASaveFile}
\title{CoTiMASaveFile}
\usage{
CoTiMASaveFile(
  activateRPB,
  activeDirectory = activeDirectory,
  SaveObject,
  FileName,
  Directory,
  silentOverwrite = FALSE
)
}
\arguments{
\item{silentOverwrite}{}
}
\value{

}
\description{
CoTiMASaveFile
}
\examples{

}
