% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaStanctArgs.R, R/data.R
\docType{data}
\name{CoTiMAStanctArgs}
\alias{CoTiMAStanctArgs}
\title{This are preset arguments}
\format{
An object of class \code{list} of length 37.
}
\usage{
CoTiMAStanctArgs

CoTiMAStanctArgs
}
\description{
This are preset arguments

object created to store standard parameters passed forward to ctStanFit
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
