% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoTiMAstudyList1}
\alias{CoTiMAstudyList1}
\title{Object created with ctmaPrep coprising 3 studies}
\format{
An object of class \code{CoTiMAFit} of length 27.
}
\usage{
CoTiMAstudyList1
}
\description{
Object created with ctmaPrep coprising 3 studies
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
