% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoTiMAstudyList_3}
\alias{CoTiMAstudyList_3}
\title{ctmaPrep-object created with 3 primary studies}
\format{
An object of class \code{CoTiMAFit} of length 28.
}
\usage{
CoTiMAstudyList_3
}
\description{
ctmaPrep-object created with 3 primary studies
}
\author{
C. Guthier, C. Dormann & J. Cortina  \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
