% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoTiMAstudyList_6}
\alias{CoTiMAstudyList_6}
\title{ctmaPrep-object created with 6 primary studies}
\format{
An object of class \code{CoTiMAFit} of length 30.
}
\usage{
CoTiMAstudyList_6
}
\description{
ctmaPrep-object created with 6 primary studies
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
