% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country201}
\alias{country201}
\title{country201 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
country201
}
\description{
country201 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
