% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country3}
\alias{country3}
\title{country3 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
country3
}
\description{
country3 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
