% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country313}
\alias{country313}
\title{country313 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
country313
}
\description{
country313 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
