% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country32}
\alias{country32}
\title{country32 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
country32
}
\description{
country32 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
