% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country4}
\alias{country4}
\title{country4 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
country4
}
\description{
country4 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
