% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaBias.R
\name{ctmaBias}
\alias{ctmaBias}
\title{ctmaBias}
\usage{
ctmaBias(
  ctmaInitFit = NULL,
  activeDirectory = NULL,
  PETPEESEalpha = 0.1,
  activateRPB = FALSE,
  digits = 4
)
}
\arguments{
\item{ctmaInitFit}{""}

\item{activeDirectory}{""}

\item{PETPEESEalpha}{""}

\item{activateRPB}{""}

\item{digits}{""}
}
\description{
ctmaBias
}
