% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaBiasOMX.R
\name{ctmaBiasOMX}
\alias{ctmaBiasOMX}
\title{ctmaBiasOMX}
\usage{
ctmaBiasOMX(
  ctmaInitFit = NULL,
  activeDirectory = NULL,
  PETPEESEalpha = 0.1,
  activateRPB = FALSE,
  digits = 4
)
}
\arguments{
\item{ctmaInitFit}{fit object created with ctmaInti containing the fitted ctsem model of each primary study}

\item{activeDirectory}{the directory where to save results (if not specified, it is taken from ctmaInitFit)}

\item{PETPEESEalpha}{# probability level (condition) below which to switch from PET to PEESE (Stanley, 2017, SPPS,p. 582, below Eq. 2; (default p = .10)}

\item{activateRPB}{if TRUE, messages (warning, finishs) could be send to smart phone (default = FALSE)}

\item{digits}{rounding (default = 4)}
}
\value{

}
\description{
Analysis of publication bias and fixed and ranom effects analysis of single drift coefficients IF OLD OMX FIT FILES ARE DETECTED
}
\examples{

}
