% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaCompFit.R
\name{ctmaCompFit}
\alias{ctmaCompFit}
\title{ctmaCompFit}
\usage{
ctmaCompFit(model1 = NULL, model2 = NULL)
}
\arguments{
\item{model1}{Model 1}

\item{model2}{Model 2}
}
\description{
Performs -2ll different tests of 2 Models (CoTiMAFit objects)
}
\examples{
minus2llDiffTest <- ctmaCompFit(CoTiMAFullInv23Fit_6,
                                CoTiMAFullInvEq23Fit_6)
summary(minus2llDiffTest)

}
