% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaCorRel.R
\name{ctmaCorRel}
\alias{ctmaCorRel}
\title{ctmaCorRel}
\usage{
ctmaCorRel(empcov = NULL, alpha = NULL)
}
\arguments{
\item{empcov}{Empirical correlation matrix}

\item{alpha}{Vector reliabilities}
}
\value{
"Empirical correlation matrix corrected for unreliability"
}
\description{
Disattenuates correlation matrix using vector of reliabilities
}
