% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaCTMultiGroupAlt.R
\name{ctmaCTMultiGroupAlt}
\alias{ctmaCTMultiGroupAlt}
\title{ctmaCTMultiGroupAlt}
\usage{
ctmaCTMultiGroupAlt(
  dat = NULL,
  groupings = NULL,
  retryattempts = 20,
  ctmodelobj = NULL,
  startValues = NULL,
  compSVMethod = "all",
  fixedmodel = NULL,
  moderators = NULL,
  moderatorLabels = NULL,
  modmodel = NULL,
  extraTries = 30,
  fastIterativeCoTiMA = FALSE,
  tryHard = FALSE,
  wtgcsv = c("prev", "best", "initial"),
  finetuneGradient = TRUE,
  checkHess = TRUE,
  method = 2,
  activateRPB = FALSE
)
}
\arguments{
\item{dat}{?}

\item{groupings}{?}

\item{retryattempts}{?}

\item{ctmodelobj}{?}

\item{startValues}{?}

\item{compSVMethod}{?}

\item{fixedmodel}{?}

\item{moderators}{?}

\item{moderatorLabels}{?}

\item{modmodel}{?}

\item{extraTries}{?}

\item{fastIterativeCoTiMA}{?}

\item{tryHard}{?}

\item{wtgcsv}{?}

\item{finetuneGradient}{?}

\item{checkHess}{?}

\item{method}{?}

\item{activateRPB}{?}
}
\value{

}
\description{
ctmaCTMultiGroupAlt
}
\examples{

}
