% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaEqual.R
\name{ctmaEqual}
\alias{ctmaEqual}
\title{ctmaEqual}
\usage{
ctmaEqual(
  ctmaInvariantFit = NULL,
  activeDirectory = NULL,
  activateRPB = FALSE,
  digits = 4,
  coresToUse = 1
)
}
\arguments{
\item{ctmaInvariantFit}{object to which a CoTiMA fit has been assigned to (i.e., what has been returned by ctmaFit). In most cases probably a model in which only (two) effects were specified with invaraintDrift.}

\item{activeDirectory}{defines another active directory than the one used in ctmaInvariantFit}

\item{activateRPB}{set to TRUE to receive push messages with CoTiMA notifications on your phone}

\item{digits}{Number of digits used for rounding (in outputs)}

\item{coresToUse}{If neg., the value is subtracted from available cores, else value = cores to use}
}
\description{
ctmaEqual
}
\examples{
# Fit a CoTiMA to all primary studies previously fitted one
# by one with the fits assigned to CoTiMAInitFit_Ex1
\dontrun{
CoTiMAEqualFit_Ex1 <- ctmaEqual(ctmaInvariantFit=CoTiMA12Fit_Ex1)
}

}
