% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaEqualMx.R
\name{ctmaEqualMx}
\alias{ctmaEqualMx}
\title{ctmaEqualMx}
\usage{
ctmaEqualMx(
  ctmaMultipleFit = NULL,
  activeDirectory = NULL,
  sourceDirectory = NULL,
  resultsFilePrefix = "CoTiMAEqualDrift",
  saveFilePrefix = "CoTiMAEqualDrift",
  activateRPB = FALSE,
  checkSingleStudyResults = FALSE,
  silentOverwrite = FALSE,
  digits = 4,
  equalDrift = NULL,
  retryattempts = 30,
  refits = 1,
  NPSOL = FALSE,
  coresToUse = c(-1),
  confidenceIntervals = FALSE,
  saveEqualDriftModelFit = saveFilePrefix
)
}
\arguments{
\item{ctmaMultipleFit}{?}

\item{activeDirectory}{?}

\item{sourceDirectory}{?}

\item{resultsFilePrefix}{?}

\item{saveFilePrefix}{?}

\item{activateRPB}{?}

\item{checkSingleStudyResults}{?}

\item{silentOverwrite}{?}

\item{digits}{?}

\item{equalDrift}{?}

\item{retryattempts}{?}

\item{refits}{?}

\item{NPSOL}{?}

\item{coresToUse}{?}

\item{confidenceIntervals}{?}

\item{saveEqualDriftModelFit}{?}
}
\value{

}
\description{
ctmaEqualMx
}
\examples{

}
