% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaExtract.R
\name{ctmaExtract}
\alias{ctmaExtract}
\title{ctmaExtract}
\usage{
ctmaExtract(
  activeDirectory = NULL,
  ctmaGenDataList = NULL,
  envir = NULL,
  useRawData = FALSE
)
}
\arguments{
\item{activeDirectory}{defines active directory where raw data files are saved. No default.}

\item{ctmaGenDataList}{list returned by \code{\link{ctmaGenData}}.}

\item{envir}{environment where objects should be extracted too. NULL by default. Typically one would use the global environment (envir = globalenv()). Has to be specified if ctmaExtract is set to TRUE.}

\item{useRawData}{if FALSE (default) empcov objects (empcov1, empcov2, etc.) are created in the global environment.
If TRUE, raw data files are created in the activeDirectory and rawData objects (rawData1, rawData2, etc.) are created in the global environment.
Note that corresponding empcov or rawData objects are deleted in the global environment if the respective other object is requested.}
}
\description{
Extracts objects required for CoTiMA from list of results returned by \code{\link{ctmaGenData}} to be used afterwards with \code{\link{ctmaPrep}}
}
