% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaFitList.R
\name{ctmaFitList}
\alias{ctmaFitList}
\title{ctmaFitList}
\usage{
ctmaFitList(...)
}
\arguments{
\item{...}{any number of CoTiMAFit objects}
}
\value{
a list that combines all objects supplied and is assigned the class 'CoTiMAFit'
}
\description{
Combines CoTiMAFit objects into a list with class CoTiMAFit to inform generic functions what to do
}
\examples{
\dontrun{
CoTiMAInitFit_3$activeDirectory <- "/Users/tmp/" # adapt!
CoTiMAFullFit_3$activeDirectory <- "/Users/tmp/" # adapt!
plot(ctmaFitList(CoTiMAInitFit_3, CoTiMAFullFit_3),
     timeUnit="Months",
     timeRange=c(1, 144, 1) )
     }

}
