% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaFittingSSMF.R
\name{ctmaFittingSSMF}
\alias{ctmaFittingSSMF}
\title{ctmaFittingSSMF}
\usage{
ctmaFittingSSMF(
  coresToUse,
  empraw,
  currentModel,
  refits,
  retryattempts,
  singleModelStartValues
)
}
\arguments{
\item{coresToUse}{?}

\item{empraw}{?}

\item{currentModel}{?}

\item{refits}{?}

\item{retryattempts}{?}

\item{singleModelStartValues}{?}
}
\value{

}
\description{
ctmaFittingSSMF
}
\examples{

}
