% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaFullMx.R
\name{ctmaFullMx}
\alias{ctmaFullMx}
\title{ctmaFullMx}
\usage{
ctmaFullMx(
  datawide_all = NULL,
  groups = NULL,
  groupsNamed = NULL,
  activeDirectory = NULL,
  resultsFilePrefix = "ctmaFullMx",
  saveFilePrefix = "ctmaFullMx",
  ctmaInitFit = NULL,
  activateRPB = NULL,
  silentOverwrite = FALSE,
  checkSingleStudyResults = TRUE,
  digits = 4,
  n.latent = NULL,
  n.studies = NULL,
  retryattempts = 1,
  refits = 1,
  NPSOL = FALSE,
  coresToUse = 1,
  numOfThreads = 1,
  fullDriftStartValues = c(),
  compSVMethod = c("mean", "fixed", "random", "rnw", "all")[2],
  useCTMultiGroupAlt = TRUE,
  confidenceIntervals = FALSE,
  saveFullDriftModelFit = saveFilePrefix
)
}
\arguments{
\item{datawide_all}{?}

\item{groups}{?}

\item{groupsNamed}{?}

\item{activeDirectory}{?}

\item{resultsFilePrefix}{?}

\item{saveFilePrefix}{?}

\item{ctmaInitFit}{?}

\item{activateRPB}{?}

\item{silentOverwrite}{?}

\item{checkSingleStudyResults}{?}

\item{digits}{?}

\item{n.latent}{?}

\item{n.studies}{?}

\item{retryattempts}{?}

\item{refits}{?}

\item{NPSOL}{?}

\item{coresToUse}{?}

\item{numOfThreads}{?}

\item{fullDriftStartValues}{?}

\item{compSVMethod}{?}

\item{useCTMultiGroupAlt}{?}

\item{confidenceIntervals}{?}

\item{saveFullDriftModelFit}{?}
}
\value{

}
\description{
ctmaFullMx
}
\examples{

}
