% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaMMtoCINT.R
\name{ctmaMMtoCINT}
\alias{ctmaMMtoCINT}
\title{ctmaMMtoCINT}
\usage{
ctmaMMtoCINT(ctmaFit = NULL)
}
\arguments{
\item{ctmaFit}{fit object created with ctmaInit or ctmaFit}
}
\value{
returns covariance of CINT-based random intercepts.
}
\description{
Compute covariance of CINT-based random intercepts obtained with a MANIFESTMEANS model specification
}
\examples{
\donttest{
RI_cov <- ctmaMMtoCINT(ctmaFit=CoTiMAFullFit_3)
print(RI_cov)
}

}
