% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaModFull.R
\name{ctmaModFull}
\alias{ctmaModFull}
\title{ctmaModFull}
\usage{
ctmaModFull(
  ctmaInitFit = NULL,
  primaryStudyList = NULL,
  cluster = NULL,
  activeDirectory = NULL,
  mod.number = 1,
  mod.type = "cont",
  mod.names = NULL,
  datalong = NULL,
  activateRPB = FALSE,
  silentOverwrite = FALSE,
  digits = 4,
  multipleDrift = c(),
  coresToUse = 1,
  CoTiMAStanctArgs = list(test = TRUE, scaleTI = TRUE, scaleTime = 1/1, scaleMod =
    TRUE, scaleLongData = FALSE, savesubjectmatrices = FALSE, verbose = 1, datalong = NA,
    ctstanmodel = NA, stanmodeltext = NA, iter = 1000, intoverstates = TRUE, binomial =
    FALSE, fit = TRUE, intoverpop = FALSE, stationary = FALSE, plot = FALSE, derrind =
    "all", optimize = TRUE, optimcontrol = list(is = F, stochastic = FALSE), nlcontrol =
    list(), nopriors = TRUE, chains = 2, cores = 1, inits = NULL, forcerecompile = FALSE,
    savescores = FALSE,      gendata = FALSE, control = list(adapt_delta = 0.8,
    adapt_window = 2, max_treedepth = 10, adapt_init_buffer = 2, stepsize = 0.001),
    verbose = 0, warmup = 500)
)
}
\arguments{
\item{ctmaInitFit}{""}

\item{primaryStudyList}{""}

\item{cluster}{""}

\item{activeDirectory}{""}

\item{mod.number}{""}

\item{mod.type}{""}

\item{mod.names}{""}

\item{datalong}{""}

\item{activateRPB}{""}

\item{silentOverwrite}{""}

\item{digits}{""}

\item{multipleDrift}{""}

\item{coresToUse}{""}

\item{CoTiMAStanctArgs}{""}

\item{type}{""}
}
\description{
ctmaModFull
}
