% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaMultipleMx.R
\name{ctmaMultipleMx}
\alias{ctmaMultipleMx}
\title{ctmaMultipleMx}
\usage{
ctmaMultipleMx(
  multipleDrift = c(),
  datawide_all = NULL,
  groups = NULL,
  groupsNamed = NULL,
  activeDirectory = NULL,
  resultsFilePrefix = "ctmaMultipleMx",
  saveFilePrefix = "ctmaMultipleMx",
  ctmaInitFit = NULL,
  activateRPB = NULL,
  silentOverwrite = FALSE,
  checkSingleStudyResults = TRUE,
  digits = 4,
  retryattempts = 1,
  refits = 1,
  NPSOL = FALSE,
  coresToUse = 1,
  numOfThreads = 1,
  fullDriftStartValues = c(),
  compSVMethod = c("mean", "fixed", "random", "rnw", "all")[2],
  useCTMultiGroupAlt = TRUE,
  confidenceIntervals = FALSE,
  saveMultipleDriftModelFit = saveMultipleDriftModelFit
)
}
\arguments{
\item{multipleDrift}{?}

\item{datawide_all}{?}

\item{groups}{?}

\item{groupsNamed}{?}

\item{activeDirectory}{?}

\item{resultsFilePrefix}{?}

\item{saveFilePrefix}{?}

\item{ctmaInitFit}{?}

\item{activateRPB}{?}

\item{silentOverwrite}{?}

\item{checkSingleStudyResults}{?}

\item{digits}{?}

\item{retryattempts}{?}

\item{refits}{?}

\item{NPSOL}{?}

\item{coresToUse}{?}

\item{numOfThreads}{?}

\item{fullDriftStartValues}{?}

\item{compSVMethod}{?}

\item{useCTMultiGroupAlt}{?}

\item{confidenceIntervals}{?}

\item{saveMultipleDriftModelFit}{?}
}
\value{

}
\description{
ctmaMultipleMx
}
\examples{

}
