% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaMultipleStanct.R
\name{ctmaMultipleStanct}
\alias{ctmaMultipleStanct}
\title{ctmaMultipleStanct}
\usage{
ctmaMultipleStanct(
  multipleDrift = c(),
  datalong_all = NULL,
  activeDirectory = NULL,
  resultsFilePrefix = "ctmaMultipleStanct",
  saveFilePrefix = "ctmaMultipleStanct",
  ctmaInitFit = NULL,
  activateRPB = NULL,
  silentOverwrite = FALSE,
  digits = 4,
  n.latent = NULL,
  n.studies = NULL,
  retryattempts = 1,
  refits = 1,
  NPSOL = FALSE,
  coresToUse = 1,
  numOfThreads = 1,
  fullDriftStartValues = c(),
  saveMultipleDriftModelFit = c(),
  CoTiMAStanctArgs = list(test = TRUE, scaleTI = TRUE, scaleTime = 1/1,
    savesubjectmatrices = FALSE, verbose = 1, datalong = NA, ctstanmodel = NA,
    stanmodeltext = NA, iter = 1000, intoverstates = TRUE, binomial = FALSE, fit = TRUE,
    intoverpop = FALSE, stationary = FALSE, plot = FALSE, derrind = "all", optimize =
    TRUE, optimcontrol = list(is = F, stochastic = FALSE), nlcontrol = list(), nopriors =
    TRUE, chains = 2, inits = NULL, forcerecompile = FALSE, savescores = FALSE, gendata =
    FALSE, control = list(adapt_delta = 0.8,      adapt_window = 2, max_treedepth = 10,
    adapt_init_buffer = 2, stepsize = 0.001), verbose = 0)
)
}
\arguments{
\item{multipleDrift}{?}

\item{datalong_all}{?}

\item{activeDirectory}{?}

\item{resultsFilePrefix}{?}

\item{saveFilePrefix}{?}

\item{ctmaInitFit}{?}

\item{activateRPB}{?}

\item{silentOverwrite}{?}

\item{digits}{?}

\item{n.latent}{?}

\item{n.studies}{?}

\item{retryattempts}{?}

\item{refits}{?}

\item{NPSOL}{?}

\item{coresToUse}{?}

\item{numOfThreads}{?}

\item{fullDriftStartValues}{?}

\item{saveMultipleDriftModelFit}{?}

\item{CoTiMAStanctArgs}{?}
}
\value{

}
\description{
ctmaMultipleStanct
}
\examples{

}
