% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaOptimizeInit.R
\name{ctmaOptimzeInit}
\alias{ctmaOptimzeInit}
\title{ctmaOptimzeInit}
\usage{
ctmaOptimzeInit(
  oldStudyList = NULL,
  activeDirectory = NULL,
  problemStudy = NULL,
  reFits = NULL,
  n.latent = NULL
)
}
\arguments{
\item{oldStudyList}{oldStudyList}

\item{activeDirectory}{activeDirectory}

\item{problemStudy}{problemStudy}

\item{reFits}{reFits}

\item{n.latent}{n.latent}
}
\description{
ctmaOptimzeInit
}
