% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaPlot.R
\name{ctmaPlot}
\alias{ctmaPlot}
\title{ctmaPlot}
\usage{
ctmaPlot(
  ctmaFitObject = list(),
  activeDirectory = NULL,
  saveFilePrefix = "ctmaPlot",
  activateRPB = FALSE,
  plotCrossEffects = TRUE,
  plotAutoEffects = TRUE,
  timeUnit = "timeUnit (not specified)",
  timeRange = c(),
  yLimitsForEffects = c(),
  mod.values = -2:2,
  mod.num = 1,
  aggregateLabel = "",
  xLabels = NULL
)
}
\arguments{
\item{ctmaFitObject}{CoTiMA Fit object}

\item{activeDirectory}{defines another active directory than the one used in ctmaInitFit}

\item{saveFilePrefix}{Prefix used for saved plots}

\item{activateRPB}{set to TRUE to receive push messages with CoTiMA notifications on your phone}

\item{plotCrossEffects}{logical}

\item{plotAutoEffects}{logical}

\item{timeUnit}{label for x-axis when plotting discrete time plots}

\item{timeRange}{time range for x-axis}

\item{yLimitsForEffects}{range for y-axis}

\item{mod.values}{moderator values that should be used for plots}

\item{mod.num}{which of possible multiple moderator effects should be plotted}

\item{aggregateLabel}{label to indicat aggregated discrete time effects}

\item{xLabels}{labes used for x-axis}
}
\description{
ctmaPlot
}
