% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaPRaw.R
\name{ctmaPRaw}
\alias{ctmaPRaw}
\title{ctmaPRaw}
\usage{
ctmaPRaw(
  empCovMat = NULL,
  empNMat = matrix(0, 0, 0),
  empN = NULL,
  studyNumber = NULL,
  empMeanVector = NULL,
  empVarVector = NULL,
  activateRPB = FALSE,
  experimental = FALSE
)
}
\arguments{
\item{empCovMat}{empirical primary study covariance matrix}

\item{empNMat}{matrix of (possibly pairwise) N}

\item{empN}{N (in case of listwise N)}

\item{studyNumber}{internal number}

\item{empMeanVector}{vector of means for all variables, usually 0}

\item{empVarVector}{vector of variances for all variables, usually 1}

\item{activateRPB}{set TRUE to receive push messages with 'CoTiMA' notifications on your phone}

\item{experimental}{set TRUE to try new pairwise N function}
}
\description{
Uses empirical correlation matrices to synthesise pseudo raw data (i.e. random data, that perfectly reproduce the correlations)
}
