% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaPub.R
\name{ctmaPub}
\alias{ctmaPub}
\title{ctmaPub}
\usage{
ctmaPub(
  getPubObj = NULL,
  primaryStudyList = NULL,
  yearsToExclude = 0,
  recency = 5,
  targetYear = NULL,
  indFUN = "sum",
  colFUN = "mean",
  addAsMod = TRUE
)
}
\arguments{
\item{getPubObj}{publication information compiled with ctmaGetPub}

\item{primaryStudyList}{vector with numbers of studies (e.g., c(1,3); requires source1 and source3 to be available)}

\item{yearsToExclude}{years to exclude from publications}

\item{recency}{years before targetYear that are considered for recency analysis}

\item{targetYear}{year (default = last year) after which publications are ignored}

\item{indFUN}{function (default = sum) how publications of each author within a collective (team) are summarized}

\item{colFUN}{function (default = mean) how publications all authors of collective (team) are summarized}

\item{addAsMod}{add to existing moderator objects (or create them) in primaryStudyList, which is part of the returned object}
}
\description{
Compute publication and citation scores for studies based on the (team of) authors' publication scores .
}
