% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaSaveFile.R
\name{ctmaSaveFile}
\alias{ctmaSaveFile}
\title{ctmaSaveFile}
\usage{
ctmaSaveFile(
  activateRPB,
  activeDirectory = activeDirectory,
  SaveObject,
  FileName,
  Directory,
  silentOverwrite = FALSE
)
}
\arguments{
\item{activateRPB}{"set to TRUE to receive push messages with CoTiMA notifications on your phone"}

\item{activeDirectory}{"Directory name"}

\item{SaveObject}{"Object to save"}

\item{FileName}{"Filename"}

\item{Directory}{"Directory to save file in"}

\item{silentOverwrite}{"override old files without asking"}
}
\description{
Internal fcuntion to save files
}
