% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaShapeRawData.R
\name{ctmaShapeRawData}
\alias{ctmaShapeRawData}
\title{ctmaShapeRawDataFiles}
\usage{
ctmaShapeRawData(
  dataFrame = NULL,
  inputDataFrameFormat = NULL,
  inputTimeFormat = "time",
  missingValues = NA,
  n.manifest = NULL,
  Tpoints = NULL,
  allInputVariablesNames = NULL,
  orderInputVariablesNames = NULL,
  targetInputVariablesNames = NULL,
  targetInputTDpredNames = NULL,
  targetInputTIpredNames = NULL,
  targetTimeVariablesNames = NULL,
  outputDataFrameFormat = "long",
  outputVariablesNames = "Y",
  outputTDpredNames = NULL,
  outputTIpredNames = NULL,
  outputTimeVariablesNames = "time",
  outputTimeFormat = "time",
  scaleTime = 1,
  mininterval = 1e-04,
  minTolDelta = NULL,
  maxTolDelta = NULL,
  negTolDelta = FALSE,
  min.val.n.Vars = 1,
  min.val.Tpoints = 1,
  experimental = FALSE
)
}
\arguments{
\item{dataFrame}{an R object containing data}

\item{inputDataFrameFormat}{"wide" or "long"}

\item{inputTimeFormat}{"time" (default) or "delta"}

\item{missingValues}{Missing value indicator, e.g., -999 or NA (default)}

\item{n.manifest}{Number of process variables (e.g, 2 in a bivariate model)}

\item{Tpoints}{Number of time points in the data frame}

\item{allInputVariablesNames}{Only required if the dataFrame (or the file header, see below) does not have column names}

\item{orderInputVariablesNames}{"names" vs "time" (e.g., names: X1, X1, X3, Y1, Y2, X3 vs X1, Y1, X2, Y2, ... )}

\item{targetInputVariablesNames}{The process variables in the dataFrame that should be used (in "names" or in "times" order)}

\item{targetInputTDpredNames}{The actual TD names, e.g, 3, or 6, or 9, ... names if Tpoints = 3. Internally, each of the 3, 6, etc represents one TDpred.}

\item{targetInputTIpredNames}{TIpred names in the dataFrame}

\item{targetTimeVariablesNames}{The time variables name in the dataFrame}

\item{outputDataFrameFormat}{"long" (default) or "wide"}

\item{outputVariablesNames}{"Y" (default), but can also be, e.g., c("X", "Y")}

\item{outputTDpredNames}{Will become "TD" if not specified}

\item{outputTIpredNames}{Will become "TI" if not specified}

\item{outputTimeVariablesNames}{"time" (default)}

\item{outputTimeFormat}{"time" (default) or "delta"}

\item{scaleTime}{A scalar that is used to multiply the time variable. Typical use is to rescale primary study time to the time scale use in other primary studies}

\item{mininterval}{param supplied to ctIntervalise. Set to lower than any possible observed measurement interval, but above 0 - this is used for filling NA values where necessary and has no impact on estimates when set in the correct range.}

\item{minTolDelta}{Set, e.g. to 1/24, to delete variables from time points that are too close (1hr; or even before) after another time point.}

\item{maxTolDelta}{Set, e.g., to 7, to delete variables from time points that are too far after another time point (e.g., 7 days, if all cases should have responed within a week)}

\item{negTolDelta}{FALSE (default) or TRUE. Delete entire cases that have at least one negative delta ('unreliable responding'; use minTolDelta to delete certain variables only)}

\item{min.val.n.Vars}{default = 1. Minimum no. of valid variables per Tpoint. Default = 1 (retaines cases with only 1 valid variable), 0 would retain cases will all variables missing (not very useful).}

\item{min.val.Tpoints}{default = 1. Minimum no. of valid Tpoints (i.e. Tpoints where min.val.n.Var.per.Tpoint is met), default = 1 retains cases with full set of valid variables at least at one single Tpoint; 2 or more retains cases which provide longitudinal information.}

\item{experimental}{FALSE (default) or TRUE. Deprecated. (Was: Shift data left if all process variables are missing at a time point (even if time stamp is available)}
}
\value{
A reshaped raw data file
}
\description{
Raw data objects are re-shaped (dealing with missing time points, wrong time intervals etc)
}
\examples{
\dontrun{
tmpData <- data.frame(matrix(c(1,  2,  1, 2,  1, 2,  11, 26, 1,
                               NA, NA, 3, NA, 3, NA, 12, 27, 1,
                               1,  2,  1, 2,  1, 2,  NA, 24, 0 ),
                          nrow=3, byrow=TRUE))
colnames(tmpData) <- c("first_T0", "second_T0", "first_T1", "second_T1",
                         "TD1_0", "TD1_1",
                        "time1", "time2", "sex")
shapedData <- ctmaShapeRawData(dataFrame=tmpData,
                               inputDataFrameFormat="wide",
                               inputTimeFormat="time",
                               n.manifest=2,
                               Tpoints=2,
                               orderInputVariablesNames="time",
                               targetInputVariablesNames=c("first_T0", "second_T0",
                                                           "first_T1", "second_T1"),
                               targetInputTDpredNames=c("TD1_0", "TD1_1"),
                               targetInputTIpredNames="sex",
                               targetTimeVariablesNames=c("time1", "time2"),
                               scaleTime=1/12,
                               maxTolDelta=1.2)
head(shapedData)
}

}
