% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaSingle.R
\name{ctmaSingle}
\alias{ctmaSingle}
\title{ctmaSingle}
\usage{
ctmaSingle(
  activeDirectory = NULL,
  sourceDirectory = NULL,
  resultsFilePrefix = "ctmaSingle",
  saveFilePrefix = "ctmaSingle",
  ctmaInitFit = NULL,
  activateRPB = FALSE,
  checkSingleStudyResults = TRUE,
  silentOverwrite = FALSE,
  digits = 4,
  single = c(),
  type = "mx",
  retryattempts = 30,
  refits = 1,
  NPSOL = FALSE,
  coresToUse = c(-1),
  fullDriftStartValues = NULL,
  compSVMethod = c("mean", "fixed", "random", "rnw", "all")[2],
  useCTMultiGroupAlt = FALSE,
  confidenceIntervals = FALSE,
  saveSingleDriftModelFit = saveFilePrefix,
  CoTiMAStanctArgs = list(test = TRUE, scaleTI = TRUE, scaleTime = 1/1,
    savesubjectmatrices = FALSE, verbose = 1, datalong = NA, ctstanmodel = NA,
    stanmodeltext = NA, iter = 1000, intoverstates = TRUE, binomial = FALSE, fit = TRUE,
    intoverpop = FALSE, stationary = FALSE, plot = FALSE, derrind = "all", optimize =
    TRUE, optimcontrol = list(is = F, stochastic = FALSE), nlcontrol = list(), nopriors =
    TRUE, chains = 2, cores = 1, inits = NULL, forcerecompile = FALSE, savescores = FALSE,
    gendata = FALSE, control = list(adapt_delta = 0.8,      adapt_window = 2,
    max_treedepth = 10, adapt_init_buffer = 2, stepsize = 0.001), verbose = 0)
)
}
\arguments{
\item{activeDirectory}{?}

\item{sourceDirectory}{?}

\item{resultsFilePrefix}{?}

\item{saveFilePrefix}{?}

\item{ctmaInitFit}{?}

\item{activateRPB}{?}

\item{checkSingleStudyResults}{?}

\item{silentOverwrite}{?}

\item{digits}{?}

\item{single}{?}

\item{type}{?}

\item{retryattempts}{?}

\item{refits}{?}

\item{NPSOL}{?}

\item{coresToUse}{?}

\item{fullDriftStartValues}{?}

\item{compSVMethod}{?}

\item{useCTMultiGroupAlt}{?}

\item{confidenceIntervals}{?}

\item{saveSingleDriftModelFit}{?}

\item{CoTiMAStanctArgs}{?}
}
\value{

}
\description{
ctmaSingle
}
\examples{

}
