% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaSingleMx.R
\name{ctmaSingleMx}
\alias{ctmaSingleMx}
\title{ctmaSingleMx}
\usage{
ctmaSingleMx(
  single = c(),
  datawide_all = NULL,
  groups = NULL,
  groupsNamed = NULL,
  activeDirectory = NULL,
  sourceDirectory = NULL,
  resultsFilePrefix = "ctmaStingleMx",
  saveFilePrefix = "ctmaStingleMx",
  ctmaInitFit = NULL,
  activateRPB = NULL,
  silentOverwrite = FALSE,
  checkSingleStudyResults = TRUE,
  digits = 4,
  retryattempts = 1,
  refits = 1,
  NPSOL = FALSE,
  coresToUse = 1,
  numOfThreads = 1,
  fullDriftStartValues = c(),
  compSVMethod = c("mean", "fixed", "random", "rnw", "all")[2],
  useCTMultiGroupAlt = TRUE,
  confidenceIntervals = FALSE,
  saveSingleDriftModelFit = saveSingleDriftModelFit
)
}
\arguments{
\item{single}{?}

\item{datawide_all}{?}

\item{groups}{?}

\item{groupsNamed}{?}

\item{activeDirectory}{?}

\item{sourceDirectory}{?}

\item{resultsFilePrefix}{?}

\item{saveFilePrefix}{?}

\item{ctmaInitFit}{?}

\item{activateRPB}{?}

\item{silentOverwrite}{?}

\item{checkSingleStudyResults}{?}

\item{digits}{?}

\item{retryattempts}{?}

\item{refits}{?}

\item{NPSOL}{?}

\item{coresToUse}{?}

\item{numOfThreads}{?}

\item{fullDriftStartValues}{?}

\item{compSVMethod}{?}

\item{useCTMultiGroupAlt}{?}

\item{confidenceIntervals}{?}

\item{saveSingleDriftModelFit}{?}
}
\value{

}
\description{
ctmaSingleMx
}
\examples{

}
