% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaStanResample.R
\name{ctmaStanResample}
\alias{ctmaStanResample}
\title{ctmaStanResample}
\usage{
ctmaStanResample(fitStanctModel = NULL, CoTiMAStanctArgs = NULL)
}
\arguments{
\item{fitStanctModel}{""}

\item{CoTiMAStanctArgs}{""}
}
\description{
ctmaStanResample
}
