% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delta_t4}
\alias{delta_t4}
\title{delta_t4 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
delta_t4
}
\description{
delta_t4 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
