% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demands1}
\alias{demands1}
\title{demands1 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
demands1
}
\description{
demands1 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
