% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demands128}
\alias{demands128}
\title{demands128 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
demands128
}
\description{
demands128 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
