% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demands18}
\alias{demands18}
\title{demands18 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
demands18
}
\description{
demands18 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
