% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demands2}
\alias{demands2}
\title{demands2 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
demands2
}
\description{
demands2 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
