% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demands201}
\alias{demands201}
\title{demands201 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
demands201
}
\description{
demands201 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
