% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demands3}
\alias{demands3}
\title{demands3 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
demands3
}
\description{
demands3 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
