% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demands313}
\alias{demands313}
\title{demands313 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
demands313
}
\description{
demands313 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
