% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demands32}
\alias{demands32}
\title{demands32 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
demands32
}
\description{
demands32 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
