% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dl_link}
\alias{dl_link}
\title{dl_link example path}
\format{
An object of class \code{character} of length 1.
}
\usage{
dl_link
}
\description{
dl_link example path
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
