% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{empcov128}
\alias{empcov128}
\title{empcov128 example matrix}
\format{
An object of class \code{list} of length 2.
}
\usage{
empcov128
}
\description{
empcov128 example matrix
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
