% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{empcov13}
\alias{empcov13}
\title{This is data to be included in my package}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 21 rows and 21 columns.
}
\usage{
empcov13
}
\description{
This is data to be included in my package
}
\references{
\url{data_blah.com}
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{data}
