% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{empcov2}
\alias{empcov2}
\title{empcov2 example matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\usage{
empcov2
}
\description{
empcov2 example matrix
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
