% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{empcov313}
\alias{empcov313}
\title{empcov313 example matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 6 columns.
}
\usage{
empcov313
}
\description{
empcov313 example matrix
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
