% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{empcov32}
\alias{empcov32}
\title{empcov32 example matrix}
\format{
An object of class \code{character} of length 4.
}
\usage{
empcov32
}
\description{
empcov32 example matrix
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
