% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{empcov6}
\alias{empcov6}
\title{This is data to be included in my package}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10 rows and 10 columns.

An object of class \code{matrix} (inherits from \code{array}) with 10 rows and 10 columns.
}
\usage{
empcov6

empcov6
}
\description{
This is data to be included in my package

This is data to be included in my package
}
\references{
\url{data_blah.com}

\url{data_blah.com}
}
\author{
My Name \email{blahblah@roxygen.org}

My Name \email{blahblah@roxygen.org}
}
\keyword{data}
