% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malePercent1}
\alias{malePercent1}
\title{malePercent1 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
malePercent1
}
\description{
malePercent1 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
