% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malePercent128}
\alias{malePercent128}
\title{malePercent128 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
malePercent128
}
\description{
malePercent128 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
