% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malePercent18}
\alias{malePercent18}
\title{malePercent18 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
malePercent18
}
\description{
malePercent18 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
