% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malePercent2}
\alias{malePercent2}
\title{malePercent2 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
malePercent2
}
\description{
malePercent2 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
