% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malePercent313}
\alias{malePercent313}
\title{malePercent313 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
malePercent313
}
\description{
malePercent313 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
