% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malePercent4}
\alias{malePercent4}
\title{malePercent4 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
malePercent4
}
\description{
malePercent4 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
