% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moderator1}
\alias{moderator1}
\title{moderator1 example vector}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
moderator1
}
\description{
moderator1 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
