% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moderator128}
\alias{moderator128}
\title{moderator128 example vector}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
moderator128
}
\description{
moderator128 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
