% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moderator18}
\alias{moderator18}
\title{moderator18 example vector}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
moderator18
}
\description{
moderator18 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
