% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moderator2}
\alias{moderator2}
\title{moderator2 example vector}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
moderator2
}
\description{
moderator2 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
