% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moderator201}
\alias{moderator201}
\title{moderator201 example vector}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
moderator201
}
\description{
moderator201 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
