% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moderator313}
\alias{moderator313}
\title{moderator313 example vector}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
moderator313
}
\description{
moderator313 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
