% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moderator32}
\alias{moderator32}
\title{moderator32 example vector}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
moderator32
}
\description{
moderator32 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
