% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moderator4}
\alias{moderator4}
\title{moderator4 example vector}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
moderator4
}
\description{
moderator4 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
