% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moderatorValues}
\alias{moderatorValues}
\title{moderatorValues example vector}
\format{
An object of class \code{list} of length 2.
}
\usage{
moderatorValues
}
\description{
moderatorValues example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
