% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupation1}
\alias{occupation1}
\title{occupation1 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
occupation1
}
\description{
occupation1 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
