% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupation128}
\alias{occupation128}
\title{occupation128 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
occupation128
}
\description{
occupation128 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
