% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupation18}
\alias{occupation18}
\title{occupation18 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
occupation18
}
\description{
occupation18 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
