% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupation201}
\alias{occupation201}
\title{occupation201 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
occupation201
}
\description{
occupation201 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
