% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupation32}
\alias{occupation32}
\title{occupation32 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
occupation32
}
\description{
occupation32 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
