% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupation4}
\alias{occupation4}
\title{occupation4 example vector}
\format{
An object of class \code{character} of length 1.
}
\usage{
occupation4
}
\description{
occupation4 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
