% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{optimFit313}
\alias{optimFit313}
\title{optimFit313 example vector}
\format{
An object of class \code{CoTiMAFit} of length 3.
}
\usage{
optimFit313
}
\description{
optimFit313 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
