% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaPRaw.R
\name{pRaw}
\alias{pRaw}
\title{ctmaPRaw}
\usage{
pRaw(
  empCovMat = NULL,
  empNMat = matrix(0, 0, 0),
  empN = NULL,
  studyNumber = NULL,
  empMeanVector = NULL,
  empVarVector = NULL,
  activateRPB = FALSE
)
}
\arguments{
\item{empCovMat}{""}

\item{empNMat}{""}

\item{empN}{""}

\item{studyNumber}{""}

\item{empMeanVector}{""}

\item{empVarVector}{""}

\item{activateRPB}{"set to TRUE to receive push messages with CoTiMA notifications on your phone"}
}
\value{

}
\description{
ctmaPRaw
}
