% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoTiMApseudoRawData.R
\name{pseudoRawData}
\alias{pseudoRawData}
\title{pseudoRawData}
\usage{
pseudoRawData(
  empCovMat = NULL,
  empNMat = matrix(0, 0, 0),
  empN = NULL,
  studyNumber = NULL,
  empMeanVector = NULL,
  empVarVector = NULL,
  activateRPB = FALSE
)
}
\arguments{
\item{empCovMat}{?}

\item{empNMat}{?}

\item{empN}{?}

\item{studyNumber}{?}

\item{empMeanVector}{?}

\item{empVarVector}{?}

\item{activateRPB}{?}
}
\value{

}
\description{
pseudoRawData
}
\examples{

}
