% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pubList_8}
\alias{pubList_8}
\title{pubList_8 example list}
\format{
An object of class \code{CoTiMAFit} of length 9.
}
\usage{
pubList_8
}
\description{
pubList_8 example list
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
