% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pubResults_6}
\alias{pubResults_6}
\title{pubResults_6 example list}
\format{
An object of class \code{CoTiMAFit} of length 6.
}
\usage{
pubResults_6
}
\description{
pubResults_6 example list
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
