% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rawData128}
\alias{rawData128}
\title{rawData128 example list}
\format{
An object of class \code{list} of length 7.
}
\usage{
rawData128
}
\description{
rawData128 example list
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
