% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recodeVariables128}
\alias{recodeVariables128}
\title{recodeVariables128 example vector}
\format{
An object of class \code{character} of length 2.
}
\usage{
recodeVariables128
}
\description{
recodeVariables128 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
