% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results128}
\alias{results128}
\title{results128 example list}
\format{
An object of class \code{list} of length 2.
}
\usage{
results128
}
\description{
results128 example list
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
