% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleSize1}
\alias{sampleSize1}
\title{sampleSize1 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
sampleSize1
}
\description{
sampleSize1 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
