% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleSize128}
\alias{sampleSize128}
\title{sampleSize128 example vector}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
sampleSize128
}
\description{
sampleSize128 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
