% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleSize18}
\alias{sampleSize18}
\title{sampleSize18 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
sampleSize18
}
\description{
sampleSize18 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
