% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleSize2}
\alias{sampleSize2}
\title{sampleSize2 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
sampleSize2
}
\description{
sampleSize2 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
