% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleSize201}
\alias{sampleSize201}
\title{sampleSize201 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
sampleSize201
}
\description{
sampleSize201 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
