% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleSize3}
\alias{sampleSize3}
\title{sampleSize3 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
sampleSize3
}
\description{
sampleSize3 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
