% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleSize313}
\alias{sampleSize313}
\title{sampleSize313 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
sampleSize313
}
\description{
sampleSize313 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
