% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleSize32}
\alias{sampleSize32}
\title{sampleSize32 example vector}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
sampleSize32
}
\description{
sampleSize32 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
