% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{source128}
\alias{source128}
\title{source128 example vector}
\format{
An object of class \code{character} of length 4.
}
\usage{
source128
}
\description{
source128 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
