% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{source18}
\alias{source18}
\title{source18 example vector}
\format{
An object of class \code{character} of length 4.
}
\usage{
source18
}
\description{
source18 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
