% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{source3}
\alias{source3}
\title{source3 example vector}
\format{
An object of class \code{character} of length 6.
}
\usage{
source3
}
\description{
source3 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
