% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{source313}
\alias{source313}
\title{source313 example vector}
\format{
An object of class \code{character} of length 4.
}
\usage{
source313
}
\description{
source313 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
