% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{source4}
\alias{source4}
\title{source4 example vector}
\format{
An object of class \code{character} of length 6.
}
\usage{
source4
}
\description{
source4 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
