% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaSummary.R
\name{summary.CoTiMAFit}
\alias{summary.CoTiMAFit}
\title{summary.CoTiMAFit}
\usage{
\method{summary}{CoTiMAFit}(object, ...)
}
\arguments{
\item{object}{one CoTiMAFit object or more as ctmaFitList(object1, object2, ...)}

\item{...}{further arguments to be passed through to summary()}
}
\description{
defins summary for CoTiMAFit objects
}
