% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targetVariables128}
\alias{targetVariables128}
\title{targetVariables128 example vector}
\format{
An object of class \code{character} of length 7.
}
\usage{
targetVariables128
}
\description{
targetVariables128 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
