% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targetVariables2}
\alias{targetVariables2}
\title{targetVariables2 example vector}
\format{
An object of class \code{character} of length 4.
}
\usage{
targetVariables2
}
\description{
targetVariables2 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
