% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targetVariables3}
\alias{targetVariables3}
\title{targetVariables3 example vector}
\format{
An object of class \code{character} of length 4.
}
\usage{
targetVariables3
}
\description{
targetVariables3 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
