% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targetVariables313}
\alias{targetVariables313}
\title{targetVariables313 example vector}
\format{
An object of class \code{character} of length 6.
}
\usage{
targetVariables313
}
\description{
targetVariables313 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
