% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variableNames18}
\alias{variableNames18}
\title{variableNames18 example vector}
\format{
An object of class \code{character} of length 4.
}
\usage{
variableNames18
}
\description{
variableNames18 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
