% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variableNames201}
\alias{variableNames201}
\title{variableNames201 example vector}
\format{
An object of class \code{character} of length 6.
}
\usage{
variableNames201
}
\description{
variableNames201 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
