% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoTiMAFullFit1}
\alias{CoTiMAFullFit1}
\title{Object created with ctmaFit using CoTiMAInitFit1 as argument}
\format{
An object of class \code{CoTiMAFit} of length 20.
}
\usage{
CoTiMAFullFit1
}
\description{
Object created with ctmaFit using CoTiMAInitFit1 as argument
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
