% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoTiMAInitFit1}
\alias{CoTiMAInitFit1}
\title{ctmaInit-object with of 3 primary studies}
\format{
An object of class \code{CoTiMAFit} of length 18.
}
\usage{
CoTiMAInitFit1
}
\description{
ctmaInit-object with of 3 primary studies
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
