% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaStanctArgs.R
\docType{data}
\name{CoTiMAStanctArgs}
\alias{CoTiMAStanctArgs}
\title{This are preset arguments}
\format{
An object of class \code{list} of length 32.
}
\usage{
CoTiMAStanctArgs
}
\description{
This are preset arguments
}
\keyword{datasets}
