% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaAllInvFit.R
\name{ctmaAllInvFit}
\alias{ctmaAllInvFit}
\title{ctmaAllInvFit}
\usage{
ctmaAllInvFit(
  ctmaInitFit = NULL,
  activeDirectory = NULL,
  activateRPB = FALSE,
  digits = 4,
  coresToUse = c(1),
  n.manifest = 0,
  indVarying = FALSE,
  scaleTime = NULL,
  optimize = TRUE,
  nopriors = TRUE,
  finishsamples = NULL,
  iter = NULL,
  chains = NULL,
  verbose = NULL,
  loadAllInvFit = c(),
  saveAllInvFit = c(),
  silentOverwrite = FALSE,
  customPar = TRUE
)
}
\arguments{
\item{ctmaInitFit}{ctmaInitFit}

\item{activeDirectory}{activeDirectory}

\item{activateRPB}{activateRPB}

\item{digits}{digits}

\item{coresToUse}{coresToUse}

\item{n.manifest}{Number of manifest variables of the model (if left empty it will assumed to be identical with n.latent).}

\item{indVarying}{Allows ct intercepts to vary at the individual level (random effects model, accounts for unobserved heterogeneity)}

\item{scaleTime}{scaleTime}

\item{optimize}{optimize}

\item{nopriors}{nopriors}

\item{finishsamples}{finishsamples}

\item{iter}{iter}

\item{chains}{chains}

\item{verbose}{verbose}

\item{loadAllInvFit}{loadAllInvFit}

\item{saveAllInvFit}{saveAllInvFit}

\item{silentOverwrite}{silentOverwrite}

\item{customPar}{logical. Leverages the first pass using priors and ensure that the drift diagonal cannott easily go too negative (could help with ctsem > 3.4)}
}
\description{
ctmaAllInvFit
}
