% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaBias.R
\name{ctmaBias}
\alias{ctmaBias}
\title{ctmaBias}
\usage{
ctmaBias(
  ctmaInitFit = NULL,
  activeDirectory = NULL,
  sourceDirectory = NULL,
  resultsFilePrefix = "CoTiMBbias",
  saveFilePrefix = "ctmaBias",
  PETPEESEalpha = 0.1,
  activateRPB = FALSE,
  silentOverwrite = FALSE,
  digits = 4
)
}
\arguments{
\item{ctmaInitFit}{?}

\item{activeDirectory}{?}

\item{sourceDirectory}{?}

\item{resultsFilePrefix}{?}

\item{saveFilePrefix}{?}

\item{PETPEESEalpha}{?}

\item{activateRPB}{?}

\item{silentOverwrite}{?}

\item{digits}{?}
}
\value{

}
\description{
ctmaBias
}
\examples{

}
