% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaBiG.R
\name{ctmaBiG}
\alias{ctmaBiG}
\title{ctmaBiG}
\usage{
ctmaBiG(
  ctmaInitFit = NULL,
  activeDirectory = NULL,
  PETPEESEalpha = 0.1,
  activateRPB = FALSE,
  digits = 4,
  zcurve = FALSE
)
}
\arguments{
\item{ctmaInitFit}{fit object created with ctmaInti containing the fitted ctsem model of each primary study}

\item{activeDirectory}{the directory where to save results (if not specified, it is taken from ctmaInitFit)}

\item{PETPEESEalpha}{# probability level (condition) below which to switch from PET to PEESE (Stanley, 2017, SPPS,p. 582, below Eq. 2; (default p = .10)}

\item{activateRPB}{if TRUE, messages (warning, finishs) could be send to smart phone (default = FALSE)}

\item{digits}{rounding (default = 4)}

\item{zcurve}{performs z-curve analysis. Could fail if too few studies (e.g. around 10) are supplied}
}
\description{
Analysis of publication bias and fixed and ranom effects analysis of single drift coefficients.
}
\examples{
\dontrun{
ctmaBiG(ctmaInitFit=studyList_Ex1)
}

}
