% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaCombPRaw.R
\name{ctmaCombPRaw}
\alias{ctmaCombPRaw}
\title{ctmaCombPRaw}
\usage{
ctmaCombPRaw(listOfStudyFits = NULL, moderatorValues = NULL)
}
\arguments{
\item{listOfStudyFits}{"Listobject of Studyfits"}

\item{moderatorValues}{"Moderators}
}
\value{

}
\description{
ctmaCombPRaw
}
