% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaCompFit.R
\name{ctmaCompFit}
\alias{ctmaCompFit}
\title{ctmaCompFit}
\usage{
ctmaCompFit(model1 = NULL, model2 = NULL)
}
\arguments{
\item{model1}{Model 1}

\item{model2}{Model 2}
}
\value{
Returns the the difference between the two -2 times LLs (Diff_Minus2LL), the associated difference in degrees of
freedom (Diff_df (= Diff_n.params)), and the probability (prob).
}
\description{
Performs log-liklihood ratio tests to compare the fit of 2 models (CoTiMAFit objects created with \code{\link{ctmaFit}}
or \code{\link{ctmaEqual}}), i.e., the difference between the two -2 times LLs between the first model and the more constrained second
model. The nested structure of the two models is assumed to be given and not checked.
}
\examples{
minus2llDiffTest <- ctmaCompFit(CoTiMAFullInv23Fit_6,
                                CoTiMAFullInvEq23Fit_6)
summary(minus2llDiffTest)

}
