% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaCorRel.R
\name{ctmaCorRel}
\alias{ctmaCorRel}
\title{ctmaCorRel}
\usage{
ctmaCorRel(empcov = NULL, alphas = NULL)
}
\arguments{
\item{empcov}{Empirical correlation matrix}

\item{alphas}{Vector reliabilities}
}
\value{
A corrected correlation matrix (corEmpcov). Corrections leading to r > 1.0 are set to 1.0.
}
\description{
Disattenuates the entries in a correlation matrix using a vector of reliabilities.
}
\examples{
empcov313new <- ctmaCorRel(empcov=empcov313, alphas=alphas313)

}
