% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaEmpCov.R
\name{ctmaEmpCov}
\alias{ctmaEmpCov}
\title{ctmaEmpCov}
\usage{
ctmaEmpCov(
  targetVariables = c(),
  recodeVariables = c(),
  combineVariables = c(),
  combineVariablesNames = c(),
  missingVariables = c(),
  nlatents = NULL,
  Tpoints = NULL,
  sampleSize = NULL,
  pairwiseN = NULL,
  empcov = NULL
)
}
\arguments{
\item{targetVariables}{""}

\item{recodeVariables}{""}

\item{combineVariables}{""}

\item{combineVariablesNames}{""}

\item{missingVariables}{""}

\item{nlatents}{""}

\item{Tpoints}{""}

\item{sampleSize}{""}

\item{pairwiseN}{""}

\item{empcov}{""}
}
\description{
ctmaEmpCov
}
