% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaEmpCov.R
\name{ctmaEmpCov}
\alias{ctmaEmpCov}
\title{ctmaEmpCov}
\usage{
ctmaEmpCov(
  targetVariables = c(),
  recodeVariables = c(),
  combineVariables = c(),
  combineVariablesNames = c(),
  missingVariables = c(),
  nlatents = NULL,
  Tpoints = NULL,
  sampleSize = NULL,
  pairwiseN = NULL,
  empcov = NULL
)
}
\arguments{
\item{targetVariables}{names of the targeted variables}

\item{recodeVariables}{which of the targeted variables require inverse coding}

\item{combineVariables}{list of vectors, which put together the targeted variables that should be used for composite variables}

\item{combineVariablesNames}{new names for combined variables - not really important}

\item{missingVariables}{missing variables}

\item{nlatents}{number of (latent) variables - actually it is the number of all variables}

\item{Tpoints}{Number of time points.}

\item{sampleSize}{sample size}

\item{pairwiseN}{matrix of same dimensions as empcov containing possible pairwiseN.}

\item{empcov}{empirical correlation matrix}
}
\description{
changes a full covariance matrix: select target variables, recode them, combine them (add), and
}
