% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaEqual.R
\name{ctmaEqual}
\alias{ctmaEqual}
\title{ctmaEqual}
\usage{
ctmaEqual(
  ctmaInvariantFit = NULL,
  activeDirectory = NULL,
  activateRPB = FALSE,
  digits = 4,
  coresToUse = 1
)
}
\arguments{
\item{ctmaInvariantFit}{object to which a CoTiMA fit has been assigned to (i.e., what has been returned by ctmaFit). In most cases probably a model in which only (two) effects were specified with invariantDrift.}

\item{activeDirectory}{defines another active directory than the one used in ctmaInvariantFit}

\item{activateRPB}{set to TRUE to receive push messages with CoTiMA notifications on your phone}

\item{digits}{Number of digits used for rounding (in outputs)}

\item{coresToUse}{If neg., the value is subtracted from available cores, else value = cores to use}
}
\description{
Test if the two or more invariant drift parameters in the CoTiMAFit object supplied are equal
}
\examples{
# Fit a CoTiMA with a set of parameters set equal that were set
# invariant in a previous model (of which the fit object is
# supplied in argument ctmaInvariantFit)
\donttest{
CoTiMAFullInv23Fit_6$activeDirectory <- "/Users/cdormann/tmp/" # adapt!
CoTiMAFullInvEq23Fit_6 <- ctmaEqual(ctmaInvariantFit=CoTiMAFullInv23Fit_6)
}

}
