% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaEqual.R
\name{ctmaEqual}
\alias{ctmaEqual}
\title{ctmaEqual}
\usage{
ctmaEqual(
  ctmaInvariantFit = NULL,
  activeDirectory = NULL,
  activateRPB = FALSE,
  digits = 4,
  equalDrift = NULL,
  coresToUse = 1,
  CoTiMAStanctArgs = list(test = TRUE, scaleTI = TRUE, scaleTime = 1/1,
    savesubjectmatrices = FALSE, verbose = 1, datalong = NA, ctstanmodel = NA,
    stanmodeltext = NA, iter = 1000, intoverstates = TRUE, binomial = FALSE, fit = TRUE,
    intoverpop = FALSE, stationary = FALSE, plot = FALSE, derrind = "all", optimize =
    TRUE, optimcontrol = list(is = F, stochastic = FALSE), nlcontrol = list(), nopriors =
    TRUE, chains = 2, cores = 1, inits = NULL, forcerecompile = FALSE, savescores =
    FALSE, gendata = FALSE, control = list(adapt_delta = 0.8,      adapt_window = 2,
    max_treedepth = 10, adapt_init_buffer = 2, stepsize = 0.001), verbose = 0)
)
}
\arguments{
\item{ctmaInvariantFit}{""}

\item{activeDirectory}{""}

\item{activateRPB}{""}

\item{digits}{""}

\item{equalDrift}{""}

\item{coresToUse}{""}

\item{CoTiMAStanctArgs}{""}
}
\description{
ctmaEqual
}
