% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaEqual.R
\name{ctmaEqual}
\alias{ctmaEqual}
\title{ctmaEqual}
\usage{
ctmaEqual(
  activeDirectory = NULL,
  sourceDirectory = NULL,
  resultsFilePrefix = "CoTiMAEqualDrift",
  saveFilePrefix = "CoTiMAEqualDrift",
  ctmaMultipleFit = NULL,
  activateRPB = FALSE,
  checkSingleStudyResults = FALSE,
  silentOverwrite = FALSE,
  digits = 4,
  equalDrift = NULL,
  retryattempts = 30,
  refits = 1,
  NPSOL = FALSE,
  coresToUse = c(-1),
  confidenceIntervals = FALSE,
  saveEqualDriftModelFit = saveFilePrefix,
  CoTiMAStanctArgs = list(test = TRUE, scaleTI = TRUE, scaleTime = 1/1,
    savesubjectmatrices = FALSE, verbose = 1, datalong = NA, ctstanmodel = NA,
    stanmodeltext = NA, iter = 1000, intoverstates = TRUE, binomial = FALSE, fit = TRUE,
    intoverpop = FALSE, stationary = FALSE, plot = FALSE, derrind = "all", optimize =
    TRUE, optimcontrol = list(is = F, stochastic = FALSE), nlcontrol = list(), nopriors =
    TRUE, chains = 2, cores = 1, inits = NULL, forcerecompile = FALSE, savescores = FALSE,
    gendata = FALSE, control = list(adapt_delta = 0.8,      adapt_window = 2,
    max_treedepth = 10, adapt_init_buffer = 2, stepsize = 0.001), verbose = 0)
)
}
\arguments{
\item{activeDirectory}{?}

\item{sourceDirectory}{?}

\item{resultsFilePrefix}{?}

\item{saveFilePrefix}{?}

\item{ctmaMultipleFit}{?}

\item{activateRPB}{?}

\item{checkSingleStudyResults}{?}

\item{silentOverwrite}{?}

\item{digits}{?}

\item{equalDrift}{?}

\item{retryattempts}{?}

\item{refits}{?}

\item{NPSOL}{?}

\item{coresToUse}{?}

\item{confidenceIntervals}{?}

\item{saveEqualDriftModelFit}{?}

\item{CoTiMAStanctArgs}{?}
}
\value{

}
\description{
ctmaEqual
}
\examples{

}
