% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaFit.R
\name{ctmaFit}
\alias{ctmaFit}
\title{ctmaFit}
\usage{
ctmaFit(
  ctmaInitFit = NULL,
  primaryStudyList = NULL,
  cluster = NULL,
  activeDirectory = NULL,
  activateRPB = FALSE,
  digits = 4,
  invariantDrift = NULL,
  drift = NULL,
  indVarying = FALSE,
  coresToUse = c(-1)
)
}
\arguments{
\item{ctmaInitFit}{""}

\item{primaryStudyList}{""}

\item{cluster}{""}

\item{activeDirectory}{""}

\item{activateRPB}{""}

\item{digits}{""}

\item{invariantDrift}{""}

\item{drift}{""}

\item{indVarying}{""}

\item{coresToUse}{""}

\item{CoTiMAStanctArgs}{""}
}
\description{
ctmaFit
}
\examples{
# CoTiMAInitFit_Ex1 <- ctmaInit(
activeDirectory = activeDirectory,
primaryStudies = studyList_Ex1,
n.latent = 2,
saveSingleStudyModelFit = c("CoTiMAInitFit_D_BO", 1:500),
silentOverwrite = TRUE)
saveRDS(CoTiMAInitFit_D_BO, file=paste0(activeDirectory, "MAIN EFFECTS/CoTiMAInitFit_D_BO.rds"))
summary(CoTiMAInitFit_D_BO)

}
