% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaFitToPrep.R
\name{ctmaFitToPrep}
\alias{ctmaFitToPrep}
\title{ctmaFitToPrep}
\usage{
ctmaFitToPrep(ctmaFitObject = NULL)
}
\arguments{
\item{ctmaFitObject}{ctmaFitObject}
}
\value{
list that could be used for fitting new CoTiMA models with \code{\link{ctmaInit}} or \code{\link{ctmaFit}}.
}
\description{
Extracts information from fitted CoTiMA objects to (re-)crearte list of primary studies originally created with \code{\link{ctmaPrep}}
}
\examples{
newStudyList <- ctmaFitToPrep(CoTiMAInitFit_3)

}
