% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaFitToPrep.R
\name{ctmaFitToPrep}
\alias{ctmaFitToPrep}
\title{ctmaFitToPrep}
\usage{
ctmaFitToPrep(ctmaFitObject = NULL)
}
\arguments{
\item{ctmaFitObject}{ctmaFitObject}
}
\description{
Extracts information from fitted CoTiMA object to (re-)crearte list of primary studies originally created with ctmaPrep
}
\examples{
newStudyList <- ctmaFitToPrep(CoTiMAInitFit_3)

}
