% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaGetPub.R
\name{ctmaGetPub}
\alias{ctmaGetPub}
\title{ctmaGetPub}
\usage{
ctmaGetPub(authorList = NULL, flush = FALSE, yearsToExclude = NULL)
}
\arguments{
\item{authorList}{list of authors and googe scholar addresses}

\item{flush}{if TRUE, the cache will be cleared and the data reloaded from Google.}

\item{yearsToExclude}{the years to be excluded (default = current year)}
}
\description{
Retrieves publication and citation information from google scholar
}
\examples{

results <- ctmaGetPub(authorList = list( c("Wilmar B.; Schaufeli",
     "https://scholar.google.de/citations?hl=en&user=w1tHcj4AAAAJ"),
     c("Maureen; Dollard",
     "user=J6oH3rgAAAAJ") ))

}
