% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaInit.R
\name{ctmaInit}
\alias{ctmaInit}
\title{ctmaInit}
\usage{
ctmaInit(
  primaryStudies = NULL,
  activeDirectory = NULL,
  activateRPB = FALSE,
  checkSingleStudyResults = TRUE,
  digits = 4,
  n.latent = NULL,
  n.manifest = 0,
  lambda = NULL,
  manifestVars = NULL,
  drift = NULL,
  indVarying = FALSE,
  saveRawData = list(),
  coresToUse = c(1),
  silentOverwrite = FALSE,
  saveSingleStudyModelFit = c(),
  loadSingleStudyModelFit = c(),
  CoTiMAStanctArgs = list(test = TRUE, scaleTI = TRUE, scaleMod = TRUE, scaleLongData =
    FALSE, scaleTime = 1/1, savesubjectmatrices = FALSE, verbose = 1, datalong = NA,
    ctstanmodel = NA, stanmodeltext = NA, iter = 1000, intoverstates = TRUE, binomial =
    FALSE, fit = TRUE, intoverpop = "auto", stationary = FALSE, plot = FALSE, derrind =
    "all", optimize = TRUE, optimcontrol = list(is = F, stochastic = FALSE, finishsamples
    = 1000), nlcontrol = list(), nopriors = TRUE, chains = 2, cores = 1, inits = NULL,
    forcerecompile = FALSE,      savescores = FALSE, gendata = FALSE, control =
    list(adapt_delta = 0.8, adapt_window = 2, max_treedepth = 10, adapt_init_buffer = 2,
    stepsize = 0.001), verbose = 0)
)
}
\arguments{
\item{primaryStudies}{list of primary study information created with ctmaPrep}

\item{activeDirectory}{Directory name}

\item{activateRPB}{set to TRUE to receive push messages with CoTiMA notifications on your phone}

\item{checkSingleStudyResults}{Displays estimates from single study ctsem models and waits for user input to continue. Useful to check estimates before the are saved.}

\item{digits}{Number of digits used for rounding (in outputs)}

\item{n.latent}{Number of latent variables of the model (hast to be specified)!}

\item{n.manifest}{Number of manifest variables of the model (if left empty it will assumed to be identical with n.latent).}

\item{lambda}{R-type matrix with pattern of fixed (=1) or free (any string) loadings.}

\item{manifestVars}{Define the error variances of the manifests with a single time point using R-type matrix with nrow=n.manifest & ncol=n.manifest.}

\item{drift}{Labels for drift effects. Have to be either of the type V1toV2 or 0 for effects to be excluded, which is usually not recommended)}

\item{indVarying}{Control for unobserved heterogeneity by having randomly (inter-individually) varying manifest means}

\item{saveRawData}{Save (created pseudo) raw date. List: saveRawData$studyNumbers, $fileName, $row.names, col.names, $sep, $dec}

\item{coresToUse}{If neg., the value is subtracted from available cores, else value = cores to use}

\item{silentOverwrite}{Override old files without asking}

\item{saveSingleStudyModelFit}{save the fit of single study ctsem models (could save a lot of time afterwards if the fit is loaded)}

\item{loadSingleStudyModelFit}{load the fit of single study ctsem models}

\item{CoTiMAStanctArgs}{CoTiMA Stanct Arguments}
}
\description{
Fits a ctsem model to a list of primary studies prepared by ctmaPrep.
#'
}
