% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaInit.R
\name{ctmaInit}
\alias{ctmaInit}
\title{ctmaInit}
\usage{
ctmaInit(
  primaryStudies = NULL,
  activeDirectory = NULL,
  loadFilePrefix = "ctmaInit",
  activateRPB = FALSE,
  checkSingleStudyResults = TRUE,
  digits = 4,
  n.latent = NULL,
  saveRawData = list(),
  coresToUse = c(1),
  silentOverwrite = FALSE,
  saveSingleStudyModelFit = c(),
  loadSingleStudyModelFit = c(),
  CoTiMAStanctArgs = list(test = TRUE, scaleTI = TRUE, scaleMod = TRUE, scaleLongData =
    FALSE, scaleTime = 1/1, savesubjectmatrices = FALSE, verbose = 1, datalong = NA,
    ctstanmodel = NA, stanmodeltext = NA, iter = 1000, intoverstates = TRUE, binomial =
    FALSE, fit = TRUE, intoverpop = FALSE, stationary = FALSE, plot = FALSE, derrind =
    "all", optimize = TRUE, optimcontrol = list(is = F, stochastic = FALSE, finishsamples
    = 1000), nlcontrol = list(), nopriors = TRUE, chains = 2, cores = 1, inits = NULL,
    forcerecompile = FALSE,      savescores = FALSE, gendata = FALSE, control =
    list(adapt_delta = 0.8, adapt_window = 2, max_treedepth = 10, adapt_init_buffer = 2,
    stepsize = 0.001), verbose = 0)
)
}
\arguments{
\item{primaryStudies}{"list of primary study information created with ctmaPrep"}

\item{activeDirectory}{"Directory name"}

\item{loadFilePrefix}{"File name prefix"}

\item{activateRPB}{"set to TRUE to receive push messages with CoTiMA notifications on your phone"}

\item{checkSingleStudyResults}{"displays estimates from single study ctsem models and waits for user inoput to continue"}

\item{digits}{"Number of digits in outputs"}

\item{n.latent}{"Number of latent variables of the model"}

\item{saveRawData}{"save (created pseudo) raw date. List: saveRawData$studyNumbers, $fileName, $row.names, col.names, $sep, $dec"}

\item{coresToUse}{"neg.if negative,value is subtracted from available cores, else value = cores to use"}

\item{silentOverwrite}{"override old files without asking"}

\item{saveSingleStudyModelFit}{"save the fit of single study ctsem models (could save a lot of time afterwards if the fit is loaded)}

\item{loadSingleStudyModelFit}{"load the fit of single study ctsem models"}

\item{CoTiMAStanctArgs}{"CoTiMA Stanct Arguments"}
}
\value{

}
\description{
ctmaInit
}
\examples{
""
}
