% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaInit.R
\name{ctmaInit}
\alias{ctmaInit}
\title{ctmaInit}
\usage{
ctmaInit(
  primaryStudies = NULL,
  activeDirectory = NULL,
  activateRPB = FALSE,
  checkSingleStudyResults = TRUE,
  digits = 4,
  n.latent = NULL,
  n.manifest = 0,
  lambda = NULL,
  manifestVars = NULL,
  drift = NULL,
  indVarying = FALSE,
  saveRawData = list(),
  coresToUse = c(1),
  silentOverwrite = FALSE,
  saveSingleStudyModelFit = c(),
  loadSingleStudyModelFit = c(),
  scaleTI = NULL,
  scaleTime = NULL,
  optimize = TRUE,
  nopriors = TRUE,
  finishsamples = NULL,
  chains = NULL,
  verbose = NULL
)
}
\arguments{
\item{primaryStudies}{list of primary study information created with ctmaPrep}

\item{activeDirectory}{defines another active directory than the one used in ctmaPrep}

\item{activateRPB}{set to TRUE to receive push messages with CoTiMA notifications on your phone}

\item{checkSingleStudyResults}{Displays estimates from single study ctsem models and waits for user input to continue. Useful to check estimates before the are saved.}

\item{digits}{Number of digits used for rounding (in outputs)}

\item{n.latent}{Number of latent variables of the model (hast to be specified)!}

\item{n.manifest}{Number of manifest variables of the model (if left empty it will assumed to be identical with n.latent).}

\item{lambda}{R-type matrix with pattern of fixed (=1) or free (any string) loadings.}

\item{manifestVars}{Define the error variances of the manifests with a single time point using R-type matrix with nrow=n.manifest & ncol=n.manifest.}

\item{drift}{Labels for drift effects. Have to be either of the type V1toV2 or 0 for effects to be excluded, which is usually not recommended)}

\item{indVarying}{Control for unobserved heterogeneity by having randomly (inter-individually) varying manifest means}

\item{saveRawData}{Save (created pseudo) raw date. List: saveRawData$studyNumbers, $fileName, $row.names, col.names, $sep, $dec}

\item{coresToUse}{If neg., the value is subtracted from available cores, else value = cores to use}

\item{silentOverwrite}{Override old files without asking}

\item{saveSingleStudyModelFit}{save the fit of single study ctsem models (could save a lot of time afterwards if the fit is loaded)}

\item{loadSingleStudyModelFit}{load the fit of single study ctsem models}

\item{scaleTI}{scale TI predictors}

\item{scaleTime}{scale time (interval) - sometimes desirable to improve fitting}

\item{optimize}{if set to FALSE, Stan’s Hamiltonian Monte Carlo sampler is used (default = TRUE = maximum a posteriori / importance sampling) .}

\item{nopriors}{if TRUE, any priors are disabled – sometimes desirable for optimization}

\item{finishsamples}{number of samples to draw (either from hessian based covariance or posterior distribution) for final results computation (default = 1000).}

\item{chains}{number of chains to sample, during HMC or post-optimization importance sampling.}

\item{verbose}{integer from 0 to 2. Higher values print more information during model fit – for debugging}
}
\description{
Fits a ctsem model to a list of primary studies prepared by ctmaPrep.
}
\examples{
# Fit a ctsem model to all primary studies summarized in
# studyList_Ex1 and save fitted models
\dontrun{
CoTiMAInitFit_Ex1 <- ctmaInit(
primaryStudies=studyList_Ex1,
n.latent=2,
saveSingleStudyModelFit=c(2, 4, 17))

summary(CoTiMAInitFit_Ex1)
}

}
