% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaLabels.R
\name{ctmaLabels}
\alias{ctmaLabels}
\title{ctmaLabels}
\usage{
ctmaLabels(
  n.latent = NULL,
  n.manifest = 0,
  lambda = NULL,
  manifestVar = NULL,
  drift = NULL,
  invariantDrift = NULL,
  moderatedDrift = NULL,
  equalDrift = NULL
)
}
\arguments{
\item{n.latent}{n.latent}

\item{n.manifest}{n.manifest}

\item{lambda}{lambda}

\item{manifestVar}{manifestVar}

\item{drift}{drift}

\item{invariantDrift}{invariantDrift}

\item{moderatedDrift}{moderatedDrift}

\item{equalDrift}{equalDrift}
}
\description{
used for consistent labeling of names and parameters
}
