% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaLabels.R
\name{ctmaLabels}
\alias{ctmaLabels}
\title{ctmaLabels}
\usage{
ctmaLabels(
  n.latent = NULL,
  n.manifest = 0,
  lambda = NULL,
  manifestVars = NULL,
  drift = NULL,
  diff = NULL,
  invariantDrift = NULL,
  moderatedDrift = NULL,
  equalDrift = NULL,
  T0means = 0,
  manifestMeans = 0
)
}
\arguments{
\item{n.latent}{n.latent}

\item{n.manifest}{n.manifest}

\item{lambda}{lambda}

\item{manifestVars}{manifestVar}

\item{drift}{drift}

\item{diff}{diffusion}

\item{invariantDrift}{invariantDrift}

\item{moderatedDrift}{moderatedDrift}

\item{equalDrift}{equalDrift}

\item{T0means}{T0means}

\item{manifestMeans}{manifestMeans}
}
\value{
returns consistently named parameters (e.g., "V1toV2") as well es their symbolic values, which are used to fix or free
parameters when fitting a 'CoTiMA' model
}
\description{
used for consistent labeling of names and parameters
}
