% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaMMtoCINT.R
\name{ctmaMMtoCINT}
\alias{ctmaMMtoCINT}
\title{ctmaMMtoCINT}
\usage{
ctmaMMtoCINT(ctmaFitObject = NULL, undoTimeScaling = FALSE, digits = 4)
}
\arguments{
\item{ctmaFitObject}{fit object created with ctmaInit or ctmaFitObject}

\item{undoTimeScaling}{if FALSE (default), results will correspond to the $randomIntercepts part of the summary of the equivalent model with indVarying="CINT"}

\item{digits}{digits used for rounding}
}
\value{
returns covariance of CINT-based random intercepts.
}
\description{
Compute covariance of CINT-based random intercepts obtained with a MANIFESTMEANS model specification
}
\examples{
\donttest{
RI_cov <- ctmaMMtoCINT(ctmaFitObject=CoTiMAFullFit_3)
print(RI_cov)
}

}
