% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaOptimizeInit.R
\name{ctmaOptimzeInit}
\alias{ctmaOptimzeInit}
\title{ctmaOptimzeInit}
\usage{
ctmaOptimzeInit(
  oldStudyList = NULL,
  activeDirectory = NULL,
  problemStudy = NULL,
  reFits = NULL,
  n.latent = NULL,
  coresToUse = c(1),
  activateRPB = FALSE,
  checkSingleStudyResults = FALSE
)
}
\arguments{
\item{oldStudyList}{oldStudyList}

\item{activeDirectory}{activeDirectory}

\item{problemStudy}{problemStudy}

\item{reFits}{reFits}

\item{n.latent}{n.latent}

\item{coresToUse}{If neg., the value is subtracted from available cores, else value = cores to use}

\item{activateRPB}{set to TRUE to receive push messages with CoTiMA notifications on your phone}

\item{checkSingleStudyResults}{Displays estimates from single study ctsem models and waits for user input to continue. Useful to check estimates before they are saved.}
}
\description{
Initial fitting a primary study reFit times to capitalize on chance for obtaining hard to find optimal fit.
}
\note{
All but one of multiple cores are used on unix-type machines for parallel fitting

During fitting, not output is generated. Be patient.
}
\examples{
\donttest{
optimFit313 <- ctmaOptimzeInit(oldStudyList=CoTiMAstudyList_3,
                               activeDirectory="/Users/cdormann/tmp/",
                               problemStudy=which(CoTiMAstudyList_3$studyNumbers == 313),
                               reFits=10,
                               n.latent=2)
summary(optimFit313)
}

}
