% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaPlot.R
\name{ctmaPlot}
\alias{ctmaPlot}
\title{ctmaPlot}
\usage{
ctmaPlot(
  ctmaFit = list(),
  activeDirectory = NULL,
  resultsFilePrefix = "ctmaPlot",
  saveFilePrefix = "ctmaPlot",
  activateRPB = FALSE,
  plotCrossEffects = TRUE,
  plotAutoEffects = TRUE,
  timeUnit = "timeUnit",
  timeRange = c(),
  yLimitsForEffects = c()
)
}
\arguments{
\item{ctmaFit}{?}

\item{activeDirectory}{?}

\item{resultsFilePrefix}{?}

\item{saveFilePrefix}{?}

\item{activateRPB}{?}

\item{plotCrossEffects}{?}

\item{plotAutoEffects}{?}

\item{timeUnit}{?}

\item{timeRange}{?}

\item{yLimitsForEffects}{?}
}
\value{

}
\description{
ctmaPlot
}
\examples{

}
