% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaPlot.R
\name{ctmaPlot}
\alias{ctmaPlot}
\title{ctmaPlot}
\usage{
ctmaPlot(
  ctmaFit = list(),
  activeDirectory = NULL,
  saveFilePrefix = "ctmaPlot",
  activateRPB = FALSE,
  plotCrossEffects = TRUE,
  plotAutoEffects = TRUE,
  timeUnit = "timeUnit (not specified)",
  timeRange = c(),
  yLimitsForEffects = c(),
  mod.values = -2:2,
  mod.num = 1,
  aggregateLabel = "SUM"
)
}
\arguments{
\item{ctmaFit}{""}

\item{activeDirectory}{""}

\item{saveFilePrefix}{""}

\item{activateRPB}{""}

\item{plotCrossEffects}{""}

\item{plotAutoEffects}{""}

\item{timeUnit}{""}

\item{timeRange}{""}

\item{yLimitsForEffects}{""}

\item{mod.values}{""}

\item{mod.num}{""}

\item{aggregateLabel}{""}
}
\description{
ctmaPlot
}
