% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaPower.R
\name{ctmaPower}
\alias{ctmaPower}
\title{ctmaPower}
\usage{
ctmaPower(
  ctmaInitFit = NULL,
  activeDirectory = NULL,
  statisticalPower = c(),
  timeRange = NULL,
  useMBESS = FALSE,
  coresToUse = 1,
  activateRPB = FALSE,
  silentOverwrite = FALSE,
  digits = 8,
  saveStatPower = "CoTiMAPower_AllResults",
  loadStatPower = NULL,
  loadAllInvFit = c(),
  saveAllInvFit = NULL,
  loadAllInvWOSingFit = c(),
  saveAllInvWOSingFit = NULL,
  skipScaling = TRUE
)
}
\arguments{
\item{ctmaInitFit}{""}

\item{activeDirectory}{""}

\item{statisticalPower}{""}

\item{timeRange}{""}

\item{useMBESS}{""}

\item{coresToUse}{""}

\item{activateRPB}{""}

\item{silentOverwrite}{""}

\item{digits}{""}

\item{saveStatPower}{""}

\item{loadStatPower}{""}

\item{loadAllInvFit}{""}

\item{saveAllInvFit}{""}

\item{loadAllInvWOSingFit}{""}

\item{saveAllInvWOSingFit}{""}

\item{skipScaling}{""}
}
\value{

}
\description{
ctmaPower
}
