% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaPower.R
\name{ctmaPower}
\alias{ctmaPower}
\title{ctmaPower}
\usage{
ctmaPower(
  ctmaInitFit = NULL,
  activeDirectory = NULL,
  resultsFilePrefix = "CoTiMBbias",
  saveFilePrefix = "ctmaBias",
  statisticalPower = c(),
  allInvStartValues = NULL,
  timeRange = NULL,
  compSVMethod = c("mean", "fixed", "random", "rnw", "all")[2],
  useMBESS = FALSE,
  type = "mx",
  retryattempts = 30,
  refits = 1,
  NPSOL = FALSE,
  coresToUse = c(-1),
  fullDriftStartValues = NULL,
  activateRPB = FALSE,
  silentOverwrite = FALSE,
  digits = 4,
  saveStatPower = "statPowerMx",
  loadStatPower = NULL,
  loadAllInvFit = c(),
  saveAllInvFit = c("ctmaAllInvMx"),
  loadAllInvWOSingFit = c(),
  saveAllInvWOSingFit = c("ctmaAllInvWOSingMx")
)
}
\arguments{
\item{ctmaInitFit}{?}

\item{activeDirectory}{?}

\item{resultsFilePrefix}{?}

\item{saveFilePrefix}{?}

\item{statisticalPower}{?}

\item{allInvStartValues}{?}

\item{timeRange}{?}

\item{compSVMethod}{?}

\item{useMBESS}{?}

\item{type}{?}

\item{retryattempts}{?}

\item{refits}{?}

\item{NPSOL}{?}

\item{coresToUse}{?}

\item{fullDriftStartValues}{?}

\item{activateRPB}{?}

\item{silentOverwrite}{?}

\item{digits}{?}

\item{saveStatPower}{?}

\item{loadStatPower}{?}

\item{loadAllInvFit}{?}

\item{saveAllInvFit}{?}

\item{loadAllInvWOSingFit}{?}

\item{saveAllInvWOSingFit}{?}
}
\value{

}
\description{
ctmaPower
}
\examples{

}
