% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaPower.R
\name{ctmaPower}
\alias{ctmaPower}
\title{ctmaPower}
\usage{
ctmaPower(
  ctmaInitFit = NULL,
  activeDirectory = NULL,
  statisticalPower = c(),
  failSafeN = NULL,
  failSafeP = NULL,
  timeRange = NULL,
  useMBESS = FALSE,
  coresToUse = 1,
  digits = 4,
  activateRPB = FALSE,
  silentOverwrite = FALSE,
  loadAllInvFit = c(),
  saveAllInvFit = NULL,
  loadAllInvWOSingFit = c(),
  saveAllInvWOSingFit = NULL,
  skipScaling = TRUE,
  useSampleFraction = NULL,
  CoTiMAStanctArgs = list(test = TRUE, scaleTI = TRUE, scaleTime = 1/1, scaleMod =
    TRUE, scaleLongData = FALSE, savesubjectmatrices = FALSE, verbose = 1, datalong = NA,
    ctstanmodel = NA, stanmodeltext = NA, iter = 1000, intoverstates = TRUE, binomial =
    FALSE, fit = TRUE, intoverpop = FALSE, stationary = FALSE, plot = FALSE, derrind =
    "all", optimize = TRUE, optimcontrol = list(is = F, stochastic = FALSE), nlcontrol =
    list(), nopriors = TRUE, chains = 2, cores = 1, inits = NULL, forcerecompile = FALSE,
    savescores = FALSE,      gendata = FALSE, control = list(adapt_delta = 0.8,
    adapt_window = 2, max_treedepth = 10, adapt_init_buffer = 2, stepsize = 0.001),
    verbose = 0, warmup = 500)
)
}
\arguments{
\item{ctmaInitFit}{""}

\item{activeDirectory}{""}

\item{statisticalPower}{""}

\item{failSafeN}{""}

\item{failSafeP}{""}

\item{timeRange}{""}

\item{useMBESS}{""}

\item{coresToUse}{""}

\item{digits}{""}

\item{activateRPB}{""}

\item{silentOverwrite}{""}

\item{loadAllInvFit}{""}

\item{saveAllInvFit}{""}

\item{loadAllInvWOSingFit}{""}

\item{saveAllInvWOSingFit}{""}

\item{skipScaling}{""}

\item{useSampleFraction}{""}

\item{CoTiMAStanctArgs}{""}
}
\description{
ctmaPower
}
