% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaPRaw.R
\name{ctmaPRaw}
\alias{ctmaPRaw}
\title{ctmaPRaw}
\usage{
ctmaPRaw(
  empCovMat = NULL,
  empNMat = matrix(0, 0, 0),
  empN = NULL,
  studyNumber = NULL,
  empMeanVector = NULL,
  empVarVector = NULL,
  activateRPB = FALSE
)
}
\arguments{
\item{empCovMat}{?}

\item{empNMat}{?}

\item{empN}{?}

\item{studyNumber}{?}

\item{empMeanVector}{?}

\item{empVarVector}{?}

\item{activateRPB}{?}
}
\value{

}
\description{
ctmaPRaw
}
\examples{

}
