% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaPrep.R
\name{ctmaPrep}
\alias{ctmaPrep}
\title{ctmaPrep}
\usage{
ctmaPrep(selectedStudies = NULL, excludedElements = NULL)
}
\arguments{
\item{selectedStudies}{?}

\item{excludedElements}{?}
}
\value{

}
\description{
ctmaPrep
}
\examples{

}
