% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaPrep.R
\name{ctmaPrep}
\alias{ctmaPrep}
\title{ctmaprep}
\usage{
ctmaPrep(selectedStudies = NULL, excludedElements = NULL, addElements = NULL)
}
\arguments{
\item{selectedStudies}{"Vector of studynumbers to include in CoTiMA"}

\item{excludedElements}{"Vector of Studyelements to exclude from CoTiMA"}

\item{addElements}{"Vector of extra Studyelements to include in CoTiMA"}
}
\value{
"List of primary studies and parameters for the following CoTiMA"
}
\description{
ctmaprep
}
