% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaSaveFile.R
\name{ctmaSaveFile}
\alias{ctmaSaveFile}
\title{ctmaSaveFile}
\usage{
ctmaSaveFile(
  activateRPB,
  activeDirectory = activeDirectory,
  SaveObject,
  FileName,
  Directory,
  silentOverwrite = FALSE
)
}
\arguments{
\item{activateRPB}{set to TRUE to receive push messages with CoTiMA notifications on your phone (FALSE by default). You have to install and configure Pushbullet to use this feature}

\item{activeDirectory}{?}

\item{SaveObject}{?}

\item{FileName}{?}

\item{Directory}{?}

\item{silentOverwrite}{?}
}
\value{

}
\description{
ctmaSaveFile
}
\examples{

}
