% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaSaveFile.R
\name{ctmaSaveFile}
\alias{ctmaSaveFile}
\title{ctmaSaveFile}
\usage{
ctmaSaveFile(
  activateRPB,
  activeDirectory = activeDirectory,
  SaveObject,
  FileName,
  Directory,
  silentOverwrite = FALSE
)
}
\arguments{
\item{activateRPB}{set TRUE to receive push messages with 'CoTiMA' notifications on your phone}

\item{activeDirectory}{directory name}

\item{SaveObject}{object to save}

\item{FileName}{filename}

\item{Directory}{directory to save file in}

\item{silentOverwrite}{override old files without asking}
}
\value{
No return value. Just saves files
}
\description{
Internal fcuntion to save files
}
