% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaScaleInits.R
\name{ctmaScaleInits}
\alias{ctmaScaleInits}
\title{ctmaScaleInits}
\usage{
ctmaScaleInits(
  CoTiMAFit = NULL,
  ctsemFit = NULL,
  newTimeScale = NULL,
  autoRefit = FALSE
)
}
\arguments{
\item{CoTiMAFit}{Fit object created with \code{\link{ctmaFit}}}

\item{ctsemFit}{Fit object created with \code{\link[ctsem]{ctStanFit}}}

\item{newTimeScale}{New Time scale}

\item{autoRefit}{Whether to automatically refit the original model using the new inits}
}
\description{
This function re-scales inits for drifts and sets all other inits to 0 (because it is too complicated to re-scale inits for diffusions).
It uses the internal transformations of \code{\link[ctsem]{ctStanFit}} (i.e., tforms) to transform the raw estimates, then re-scale them, and finally use the inverse of tform to supply raw estimates as inits.
}
