% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaStanResample.R
\name{ctmaStanResample}
\alias{ctmaStanResample}
\title{ctmaStanResample}
\usage{
ctmaStanResample(ctmaFittedModel = NULL, nsamples = 25, overallSamples = 500)
}
\arguments{
\item{ctmaFittedModel}{a model fitted ith CoTiMA R package, usually with few 'finishsamples' to prevent memory exhaustion}

\item{nsamples}{sample size per rund}

\item{overallSamples}{overall samples isze to be achieved}
}
\description{
Re-sample from a fitted stanct model to achieve desired number of finishsamples (could be useful to prevent exhausted memory)
}
