% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaStanResample.R
\name{ctmaStanResample}
\alias{ctmaStanResample}
\title{ctmaStanResample}
\usage{
ctmaStanResample(ctmaFittedModel = NULL, nsamples = 25, overallSamples = 500)
}
\arguments{
\item{ctmaFittedModel}{a 'CoTiMA' fit object, usually with few 'finishsamples' to prevent memory exhaustion}

\item{nsamples}{sample size per run}

\item{overallSamples}{overall samples size to be achieved}
}
\value{
returns a CoTiMA fit object with an increased number of finish samples
}
\description{
re-sample from a fitted stanct model to achieve desired number of finishsamples (could be useful to prevent exhausted memory)
}
