% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pairwiseN128}
\alias{pairwiseN128}
\title{pairwiseN128 example vector}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 9 rows and 9 columns.
}
\usage{
pairwiseN128
}
\description{
pairwiseN128 example vector
}
\author{
C. Dormann & M. Homburg \email{CoTiMA@uni-mainz.org}
}
\keyword{data}
