% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmaPlotContainer.R
\name{plot.CoTiMAFit}
\alias{plot.CoTiMAFit}
\title{plot.CoTiMAFit}
\usage{
\method{plot}{CoTiMAFit}(x, ...)
}
\arguments{
\item{x}{líst}

\item{...}{further arguments to be passed through to summary()}
}
\value{
returns a call to 'ctmaPlot', which is used to plot CoTiMA fit objects
}
\description{
call \code{\link{ctmaPlot}} if a CoTiMAFit object is supplied to plot()
}
